@@ -1,21 +1,37 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <Version>1.0.3.4</Version>
+    <TargetFramework Condition="'$(TargetFrameworkOverride)' == ''">netcoreapp3.1</TargetFramework>
+    <TargetFramework Condition="'$(TargetFrameworkOverride)' != ''">$(TargetFrameworkOverride)</TargetFramework>
+    <Version>2.1.3</Version>
 	  <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\NBXplorer.xml</DocumentationFile>
 	  <NoWarn>1701;1702;1705;1591;CS1591</NoWarn>
-	  <LangVersion>7.3</LangVersion>
+	  <LangVersion>8.0</LangVersion>
+    <RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
   </PropertyGroup>
   <ItemGroup>
-	<PackageReference Include="DBreeze" Version="1.87.0" />
-	  <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.4" />
-	  <PackageReference Include="Microsoft.Azure.ServiceBus" Version="3.1.0" />
-	<PackageReference Include="NicolasDorier.CommandLine" Version="1.0.0.2" />
-	<PackageReference Include="NicolasDorier.CommandLine.Configuration" Version="1.0.0.3" />
-    <PackageReference Include="System.Xml.XmlSerializer" Version="4.3.0" />
+    <None Remove="banlist.cli.txt" />
   </ItemGroup>
   <ItemGroup>
+    <EmbeddedResource Include="banlist.cli.txt" />
+  </ItemGroup>
+  <PropertyGroup Condition="'$(TargetFramework)' == 'netcoreapp2.1'">
+    <DefineConstants>$(DefineConstants);NETCOREAPP21</DefineConstants>
+  </PropertyGroup>
+  <ItemGroup Condition="'$(TargetFramework)' != 'netcoreapp2.1'">
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.0" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp2.1'">
+	  <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.9" />
+    <Compile Remove="HealthChecks\**"></Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.ServiceBus" Version="4.0.0" />
+    <PackageReference Include="RabbitMQ.Client" Version="5.1.0" />
+    <PackageReference Include="DBriize" Version="1.0.1.3" />
+	  <PackageReference Include="NicolasDorier.CommandLine" Version="1.0.0.2" />
+	  <PackageReference Include="NicolasDorier.CommandLine.Configuration" Version="1.0.0.3" />
     <ProjectReference Include="..\NBXplorer.Client\NBXplorer.Client.csproj" />
   </ItemGroup>
   <ItemGroup>
