@@ -18,17 +18,21 @@ using NBXplorer.Filters;
 using NBXplorer.Logging;
 using Microsoft.AspNetCore.Authentication;
 using NBXplorer.Authentication;
+#if NETCOREAPP21
+using IWebHostEnvironment = Microsoft.AspNetCore.Hosting.IHostingEnvironment;
+#else
+using Newtonsoft.Json.Serialization;
+using Microsoft.Extensions.Hosting;
+#endif
 
 namespace NBXplorer
 {
 	public class Startup
 	{
-		public Startup(IConfiguration conf, IHostingEnvironment env)
+		public Startup(IConfiguration conf)
 		{
 			Configuration = conf;
-			_Env = env;
 		}
-		IHostingEnvironment _Env;
 		public IConfiguration Configuration
 		{
 			get;
@@ -36,17 +40,34 @@ namespace NBXplorer
 
 		public void ConfigureServices(IServiceCollection services)
 		{
+			services.AddHttpClient();
+			services.AddHttpClient(nameof(RPCClientProvider), httpClient =>
+			{
+				httpClient.Timeout = System.Threading.Timeout.InfiniteTimeSpan;
+			});
 			services.AddNBXplorer();
 			services.ConfigureNBxplorer(Configuration);
-			services.AddMvcCore()
-				.AddJsonFormatters()
-				.AddAuthorization()
-				.AddFormatterMappings();
+			var builder = services.AddMvcCore();
+#if NETCOREAPP21
+			builder.AddJsonFormatters();
+#else
+			services.AddHealthChecks().AddCheck<HealthChecks.NodesHealthCheck>("NodesHealthCheck");
+			builder.AddNewtonsoftJson(options =>
+			{
+				options.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
+				new Serializer(null).ConfigureSerializer(options.SerializerSettings);
+			});
+#endif
+			builder.AddMvcOptions(o => o.InputFormatters.Add(new NoContentTypeInputFormatter()))
+			.AddAuthorization()
+			.AddFormatterMappings();
 			services.AddAuthentication("Basic")
 				.AddNBXplorerAuthentication();
 		}
 
-		public void Configure(IApplicationBuilder app, IServiceProvider prov, IHostingEnvironment env, ILoggerFactory loggerFactory, IServiceProvider serviceProvider,
+		public void Configure(IApplicationBuilder app, IServiceProvider prov,
+			IWebHostEnvironment env,
+			ILoggerFactory loggerFactory, IServiceProvider serviceProvider,
 			CookieRepository cookieRepository)
 		{
 			cookieRepository.Initialize();
@@ -54,11 +75,29 @@ namespace NBXplorer
 			{
 				app.UseDeveloperExceptionPage();
 			}
-
 			Logs.Configure(loggerFactory);
+
+#if !NETCOREAPP21
+			app.UseRouting();
+#endif
 			app.UseAuthentication();
+#if !NETCOREAPP21
+			app.UseAuthorization();
+#endif
 			app.UseWebSockets();
+			//app.UseMiddleware<LogAllRequestsMiddleware>();
+#if NETCOREAPP21
 			app.UseMvc();
+#else
+			app.UseEndpoints(endpoints =>
+			{
+				endpoints.MapHealthChecks("health", new Microsoft.AspNetCore.Diagnostics.HealthChecks.HealthCheckOptions()
+				{
+					ResponseWriter = HealthChecks.HealthCheckWriters.WriteJSON
+				});
+				endpoints.MapControllers();
+			});
+#endif
 		}
 	}
 }
