@@ -9,8 +9,22 @@ namespace NBXplorer.Models
 {
 	public class KeyPathInformation
 	{
-		public TrackedSource TrackedSource { get; set; }
+		public KeyPathInformation()
+		{
 
+		}
+		
+		public KeyPathInformation(Derivation derivation, DerivationSchemeTrackedSource derivationStrategy, DerivationFeature feature, KeyPath keyPath, NBXplorerNetwork network)
+		{
+			ScriptPubKey = derivation.ScriptPubKey;
+			Redeem = derivation.Redeem;
+			TrackedSource = derivationStrategy;
+			DerivationStrategy = derivationStrategy.DerivationStrategy;
+			Feature = feature;
+			KeyPath = keyPath;
+			Address =  network.CreateAddress(derivationStrategy.DerivationStrategy, keyPath, ScriptPubKey);
+		}
+		public TrackedSource TrackedSource { get; set; }
 		[JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
 		public DerivationFeature Feature
 		{
@@ -30,9 +44,16 @@ namespace NBXplorer.Models
 		{
 			get; set;
 		}
+		[JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
+		public BitcoinAddress Address { get; set; }
+		[JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
 		public Script Redeem
 		{
 			get; set;
+		}
+		public int GetIndex()
+		{
+			return (int)KeyPath.Indexes[KeyPath.Indexes.Length - 1];
 		}
 	}
 }
