@@ -2,6 +2,6 @@
 {
 	public class NewBlockEventRequest : NewEventBase
 	{
-
+		public override string EventType => "subscribeblock";
 	}
 }
