@@ -60,25 +60,19 @@ namespace NBXplorer.Models
 			}
 		}
 
-		public bool HasChanges
-		{
-			get
-			{
-				return Confirmed.HasChanges || Unconfirmed.HasChanges;
-			}
-		}
-
 		public Coin[] GetUnspentCoins(bool excludeUnconfirmedUTXOs = false)
 		{
-			if (Confirmed.KnownBookmark != null || Unconfirmed.KnownBookmark != null)
-				throw new InvalidOperationException("This UTXOChanges is partial, it is calculate the unspent coins");
-			return GetUnspentUTXOs(excludeUnconfirmedUTXOs).Select(c => c.AsCoin(DerivationStrategy)).ToArray();
+			return GetUnspentCoins(excludeUnconfirmedUTXOs ? 1 : 0);
 		}
-
-		public UTXO[] GetUnspentUTXOs(bool excludeUnconfirmedUTXOs = false)
+		public Coin[] GetUnspentCoins(int minConfirmations)
 		{
+			return GetUnspentUTXOs(minConfirmations).Select(c => c.AsCoin(DerivationStrategy)).ToArray();
+		}
+		public UTXO[] GetUnspentUTXOs(int minConf)
+		{
+			var excludeUnconfirmedUTXOs = minConf > 0;
 			Dictionary<OutPoint, UTXO> received = new Dictionary<OutPoint, UTXO>();
-			foreach (var utxo in Confirmed.UTXOs.Concat(excludeUnconfirmedUTXOs ? (IEnumerable<UTXO>)Array.Empty<UTXO>() : Unconfirmed.UTXOs))
+			foreach (var utxo in Confirmed.UTXOs.Where(u => u.Confirmations >= minConf).Concat(excludeUnconfirmedUTXOs ? (IEnumerable<UTXO>)Array.Empty<UTXO>() : Unconfirmed.UTXOs))
 			{
 				received.TryAdd(utxo.Outpoint, utxo);
 			}
@@ -88,6 +82,10 @@ namespace NBXplorer.Models
 			}
 			return received.Values.ToArray();
 		}
+		public UTXO[] GetUnspentUTXOs(bool excludeUnconfirmedUTXOs = false)
+		{
+			return GetUnspentUTXOs(excludeUnconfirmedUTXOs ? 1 : 0);
+		}
 
 		public Key[] GetKeys(ExtKey extKey, bool excludeUnconfirmedUTXOs = false)
 		{
@@ -96,32 +94,6 @@ namespace NBXplorer.Models
 	}
 	public class UTXOChange
 	{
-
-		Bookmark _KnownBookmark;
-		public Bookmark KnownBookmark
-		{
-			get
-			{
-				return _KnownBookmark;
-			}
-			set
-			{
-				_KnownBookmark = value;
-			}
-		}
-
-		Bookmark _Bookmark = null;
-		public Bookmark Bookmark
-		{
-			get
-			{
-				return _Bookmark;
-			}
-			set
-			{
-				_Bookmark = value;
-			}
-		}
 
 		List<UTXO> _UTXOs = new List<UTXO>();
 		public List<UTXO> UTXOs
@@ -148,14 +120,6 @@ namespace NBXplorer.Models
 				_SpentOutpoints = value;
 			}
 		}
-
-		public bool HasChanges
-		{
-			get
-			{
-				return KnownBookmark != Bookmark || UTXOs.Count != 0 || SpentOutpoints.Count != 0;
-			}
-		}
 	}
 
 	public class UTXO
@@ -165,14 +129,17 @@ namespace NBXplorer.Models
 
 		}
 
-		public UTXO(Coin coin)
+		public UTXO(ICoin coin)
 		{
 			Outpoint = coin.Outpoint;
-			Value = coin.TxOut.Value;
+			Index = (int)coin.Outpoint.N;
+			TransactionHash = coin.Outpoint.Hash;
+			Value = coin.Amount;
 			ScriptPubKey = coin.TxOut.ScriptPubKey;
 		}
 
 		[JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
+		[JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
 		public DerivationFeature? Feature
 		{
 			get; set;
@@ -185,16 +152,20 @@ namespace NBXplorer.Models
 
 		public Coin AsCoin(DerivationStrategy.DerivationStrategyBase derivationStrategy)
 		{
-			var coin = new Coin(Outpoint, new TxOut(Value, ScriptPubKey));
-			if (derivationStrategy != null)
+			if (Value is Money v)
 			{
-				var derivation = derivationStrategy.Derive(KeyPath);
-				if (derivation.ScriptPubKey != coin.ScriptPubKey)
-					throw new InvalidOperationException($"This Derivation Strategy does not own this coin");
-				if (derivation.Redeem != null)
-					coin = coin.ToScriptCoin(derivation.Redeem);
+				var coin = new Coin(Outpoint, new TxOut(v, ScriptPubKey));
+				if (derivationStrategy != null)
+				{
+					var derivation = derivationStrategy.GetDerivation(KeyPath);
+					if (derivation.ScriptPubKey != coin.ScriptPubKey)
+						throw new InvalidOperationException($"This Derivation Strategy does not own this coin");
+					if (derivation.Redeem != null)
+						coin = coin.ToScriptCoin(derivation.Redeem);
+				}
+				return coin;
 			}
-			return coin;
+			return null;
 		}
 
 		OutPoint _Outpoint = new OutPoint();
@@ -210,7 +181,8 @@ namespace NBXplorer.Models
 			}
 		}
 
-
+		public int Index { get; set; }
+		public uint256 TransactionHash { get; set; }
 
 		Script _ScriptPubKey;
 		public Script ScriptPubKey
@@ -226,8 +198,8 @@ namespace NBXplorer.Models
 		}
 
 
-		Money _Value;
-		public Money Value
+		IMoney _Value;
+		public IMoney Value
 		{
 			get
 			{
