@@ -6,14 +6,65 @@ using System.Text;
 using System.Linq;
 using NBXplorer.DerivationStrategy;
 using NBitcoin.RPC;
+using System.Threading;
 
 namespace NBXplorer.Tests
 {
 	public static class Extensions
 	{
+		public static void WaitForTransaction(this LongPollingNotificationSession session, DerivationStrategyBase derivationStrategy, uint256 txId)
+		{
+			session.WaitForTransaction(TrackedSource.Create(derivationStrategy), txId);
+		}
+		public static void WaitForTransaction(this LongPollingNotificationSession session, BitcoinAddress address, uint256 txId)
+		{
+			session.WaitForTransaction(TrackedSource.Create(address), txId);
+		}
+		public static void WaitForTransaction(this LongPollingNotificationSession session, TrackedSource trackedSource, uint256 txId)
+		{
+			using (var cts = new CancellationTokenSource(TimeSpan.FromSeconds(10)))
+			{
+				while (true)
+				{
+					if (session.NextEvent(cts.Token) is NewTransactionEvent evts)
+					{
+						if (evts.TrackedSource == trackedSource && evts.TransactionData.TransactionHash == txId)
+						{
+							break;
+						}
+					}
+				}
+			}
+		}
+
+		public static void WaitForBlocks(this LongPollingNotificationSession session, params uint256[] txIds)
+		{
+			if (txIds == null || txIds.Length == 0)
+				return;
+			HashSet<uint256> txidsSet = new HashSet<uint256>(txIds);
+			using (var cts = new CancellationTokenSource(TimeSpan.FromSeconds(20)))
+			{
+				while (true)
+				{
+					if (session.NextEvent(cts.Token) is NewBlockEvent evts)
+					{
+						txidsSet.Remove(evts.Hash);
+						if (txidsSet.Count == 0)
+							break;
+					}
+				}
+			}
+		}
+
+		public static DerivationLine GetLineFor(this DerivationStrategyBase strategy, DerivationFeature feature)
+		{
+			return strategy.GetLineFor(KeyPathTemplates.Default.GetKeyPathTemplate(feature));
+		}
+
+		static BitcoinAddress Dummy = new Key().PubKey.GetAddress(ScriptPubKeyType.Legacy, Network.Main);
 		public static KeyPathInformation GetKeyInformation(this Repository repo, Script script)
 		{
-			return repo.GetKeyInformations(new Script[] { script })[script].SingleOrDefault();
+			return repo.GetKeyInformations(new Script[] { script }).GetAwaiter().GetResult()[script].SingleOrDefault();
 		}
 
 		public static uint256[] EnsureGenerate(this RPCClient client, int blockCount)
@@ -21,26 +72,10 @@ namespace NBXplorer.Tests
 			return client.EnsureGenerateAsync(blockCount).GetAwaiter().GetResult();
 		}
 
-		public static Transaction SignRawTransaction(this ServerTester tester, Transaction transaction)
+		public static RPCClient WithCapabilitiesOf(this RPCClient client, RPCClient target)
 		{
-			return SignRawTransaction(tester.RPC, transaction, tester);
-		}
-
-		public static Transaction SignRawTransaction(this RPCClient client, Transaction transaction, ServerTester tester)
-		{
-			if (tester.RPCSupportSignRawTransaction)
-			{
-#pragma warning disable CS0618 // Type or member is obsolete
-				return client.SignRawTransaction(transaction);
-#pragma warning restore CS0618 // Type or member is obsolete
-			}
-			else
-			{
-				return client.SignRawTransactionWithWallet(new SignRawTransactionRequest()
-				{
-					Transaction = transaction
-				}).SignedTransaction;
-			}
+			client.Capabilities = target.Capabilities;
+			return client;
 		}
 	}
 
