@@ -33,15 +33,19 @@ namespace NBXplorer.Tests
 		{
 			return new ServerTester(caller);
 		}
+		public static ServerTester CreateNoAutoStart([CallerMemberNameAttribute]string caller = null)
+		{
+			return new ServerTester(caller, false);
+		}
 
 		public void Dispose()
 		{
-			if(Host != null)
+			if (Host != null)
 			{
 				Host.Dispose();
 				Host = null;
 			}
-			if(NodeBuilder != null)
+			if (NodeBuilder != null)
 			{
 				NodeBuilder.Dispose();
 				NodeBuilder = null;
@@ -55,85 +59,39 @@ namespace NBXplorer.Tests
 			get; set;
 		}
 
-		public ServerTester(string directory)
+		public ServerTester(string directory, bool autoStart = true)
 		{
 			SetEnvironment();
+			var rootTestData = "TestData";
+			directory = Path.Combine(rootTestData, directory);
+			_Directory = directory;
+			if (!Directory.Exists(rootTestData))
+				Directory.CreateDirectory(rootTestData);
+			if (autoStart)
+				Start();
+		}
+
+		public void Start()
+		{
 			try
 			{
-				var rootTestData = "TestData";
-				directory = Path.Combine(rootTestData, directory);
-				_Directory = directory;
-				if(!Directory.Exists(rootTestData))
-					Directory.CreateDirectory(rootTestData);
-
 				var cryptoSettings = new NBXplorerNetworkProvider(NetworkType.Regtest).GetFromCryptoCode(CryptoCode);
-				NodeBuilder = NodeBuilder.Create(nodeDownloadData, Network, directory);
+				NodeBuilder = NodeBuilder.Create(nodeDownloadData, Network, _Directory);
 
-
-				User1 = NodeBuilder.CreateNode();
-				User2 = NodeBuilder.CreateNode();
 				Explorer = NodeBuilder.CreateNode();
-				foreach(var node in NodeBuilder.Nodes)
+				Explorer.ConfigParameters.Add("txindex", "1");
+				foreach (var node in NodeBuilder.Nodes)
 				{
 					node.WhiteBind = true;
 					node.CookieAuth = cryptoSettings.SupportCookieAuthentication;
 				}
 				NodeBuilder.StartAll();
+				Explorer.CreateRPCClient().EnsureGenerate(Network.Consensus.CoinbaseMaturity + 1);
 
-				User1.CreateRPCClient().Generate(1);
-				User1.Sync(Explorer, true);
-				Explorer.CreateRPCClient().Generate(1);
-				Explorer.Sync(User2, true);
-				User2.CreateRPCClient().EnsureGenerate(Network.Consensus.CoinbaseMaturity + 1);
-				User1.Sync(User2, true);
-
-				var port = CustomServer.FreeTcpPort();
-				var datadir = Path.Combine(directory, "explorer");
-				DeleteRecursivelyWithMagicDust(datadir);
-				List<(string key, string value)> keyValues = new List<(string key, string value)>();
-				keyValues.Add(("conf", Path.Combine(directory, "explorer", "settings.config")));
-				keyValues.Add(("datadir", datadir));
-				keyValues.Add(("port", port.ToString()));
-				keyValues.Add(("network", "regtest"));
-				keyValues.Add(("chains", CryptoCode.ToLowerInvariant()));
-				keyValues.Add(("verbose", "1"));
-				keyValues.Add(($"{CryptoCode.ToLowerInvariant()}rpcauth", Explorer.GetRPCAuth()));
-				keyValues.Add(($"{CryptoCode.ToLowerInvariant()}rpcurl", Explorer.CreateRPCClient().Address.AbsoluteUri));
-				keyValues.Add(("cachechain", "0"));
-				keyValues.Add(("rpcnotest", "1"));
-				keyValues.Add(("mingapsize", "2"));
-				keyValues.Add(("maxgapsize", "4"));
-				keyValues.Add(($"{CryptoCode.ToLowerInvariant()}startheight", Explorer.CreateRPCClient().GetBlockCount().ToString()));
-				keyValues.Add(($"{CryptoCode.ToLowerInvariant()}nodeendpoint", $"{Explorer.Endpoint.Address}:{Explorer.Endpoint.Port}"));
-				keyValues.Add(("asbcnstr", AzureServiceBusTestConfig.ConnectionString));
-				keyValues.Add(("asbblockq", AzureServiceBusTestConfig.NewBlockQueue));
-				keyValues.Add(("asbtranq", AzureServiceBusTestConfig.NewTransactionQueue));
-				keyValues.Add(("asbblockt", AzureServiceBusTestConfig.NewBlockTopic));
-				keyValues.Add(("asbtrant", AzureServiceBusTestConfig.NewTransactionTopic));
-
-				var args = keyValues.SelectMany(kv => new[] { $"--{kv.key}", kv.value }).ToArray();
-				Host = new WebHostBuilder()
-					.UseConfiguration(new DefaultConfiguration().CreateConfiguration(args))
-					.UseKestrel()
-					.ConfigureLogging(l =>
-					{
-						l.SetMinimumLevel(LogLevel.Information)
-							.AddFilter("Microsoft", LogLevel.Error)
-							.AddFilter("Hangfire", LogLevel.Error)
-							.AddFilter("NBXplorer.Authentication.BasicAuthenticationHandler", LogLevel.Critical)
-							.AddProvider(Logs.LogProvider);
-					})
-					.UseStartup<Startup>()
-					.Build();
-
-				RPC = ((RPCClientProvider)Host.Services.GetService(typeof(RPCClientProvider))).GetRPCClient(CryptoCode);
-				var nbxnetwork = ((NBXplorerNetworkProvider)Host.Services.GetService(typeof(NBXplorerNetworkProvider))).GetFromCryptoCode(CryptoCode);
-				Network = nbxnetwork.NBitcoinNetwork;
-				var conf = (ExplorerConfiguration)Host.Services.GetService(typeof(ExplorerConfiguration));
-				Host.Start();
-
-				_Client = new ExplorerClient(nbxnetwork, Address);
-				_Client.SetCookieAuth(Path.Combine(conf.DataDir, ".cookie"));
+				datadir = Path.Combine(_Directory, "explorer");
+				DeleteFolderRecursive(datadir);
+				StartNBXplorer();
+				this.Client.WaitServerStarted();
 			}
 			catch
 			{
@@ -142,6 +100,76 @@ namespace NBXplorer.Tests
 			}
 		}
 
+		public bool UseRabbitMQ { get; set; } = false;
+		private void StartNBXplorer()
+		{
+			var port = CustomServer.FreeTcpPort();
+			List<(string key, string value)> keyValues = new List<(string key, string value)>();
+			keyValues.Add(("conf", Path.Combine(datadir, "settings.config")));
+			keyValues.Add(("datadir", datadir));
+			keyValues.Add(("port", port.ToString()));
+			keyValues.Add(("network", "regtest"));
+			keyValues.Add(("chains", CryptoCode.ToLowerInvariant()));
+			keyValues.Add(("verbose", "1"));
+			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}rpcauth", Explorer.GetRPCAuth()));
+			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}rpcurl", Explorer.CreateRPCClient().Address.AbsoluteUri));
+			keyValues.Add(("cachechain", "0"));
+			keyValues.Add(("rpcnotest", "1"));
+			keyValues.Add(("mingapsize", "3"));
+			keyValues.Add(("maxgapsize", "8"));
+			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}startheight", Explorer.CreateRPCClient().GetBlockCount().ToString()));
+			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}nodeendpoint", $"{Explorer.Endpoint.Address}:{Explorer.Endpoint.Port}"));
+			keyValues.Add(("asbcnstr", AzureServiceBusTestConfig.ConnectionString));
+			keyValues.Add(("asbblockq", AzureServiceBusTestConfig.NewBlockQueue));
+			keyValues.Add(("asbtranq", AzureServiceBusTestConfig.NewTransactionQueue));
+			keyValues.Add(("asbblockt", AzureServiceBusTestConfig.NewBlockTopic));
+			keyValues.Add(("asbtrant", AzureServiceBusTestConfig.NewTransactionTopic));
+			if (UseRabbitMQ)
+			{
+				keyValues.Add(("rmqhost", RabbitMqTestConfig.RabbitMqHostName));
+				keyValues.Add(("rmqvirtual", RabbitMqTestConfig.RabbitMqVirtualHost));
+				keyValues.Add(("rmquser", RabbitMqTestConfig.RabbitMqUsername));
+				keyValues.Add(("rmqpass", RabbitMqTestConfig.RabbitMqPassword));
+				keyValues.Add(("rmqtranex", RabbitMqTestConfig.RabbitMqTransactionExchange));
+				keyValues.Add(("rmqblockex", RabbitMqTestConfig.RabbitMqBlockExchange));
+			}
+			var args = keyValues.SelectMany(kv => new[] { $"--{kv.key}", kv.value }
+			.Concat(new[] { $"--{CryptoCode.ToLowerInvariant()}hastxindex" })).ToArray();
+			Host = new WebHostBuilder()
+				.UseConfiguration(new DefaultConfiguration().CreateConfiguration(args))
+				.UseKestrel()
+				.ConfigureLogging(l =>
+				{
+					l.SetMinimumLevel(LogLevel.Information)
+						.AddFilter("System.Net.Http.HttpClient", LogLevel.Error)
+						.AddFilter("Microsoft", LogLevel.Error)
+						.AddFilter("Hangfire", LogLevel.Error)
+						.AddFilter("NBXplorer.Authentication.BasicAuthenticationHandler", LogLevel.Critical)
+						.AddProvider(Logs.LogProvider);
+				})
+				.UseStartup<Startup>()
+				.Build();
+
+			RPC = ((RPCClientProvider)Host.Services.GetService(typeof(RPCClientProvider))).GetRPCClient(CryptoCode);
+			NBXplorerNetwork = ((NBXplorerNetworkProvider)Host.Services.GetService(typeof(NBXplorerNetworkProvider))).GetFromCryptoCode(CryptoCode);
+			var conf = (ExplorerConfiguration)Host.Services.GetService(typeof(ExplorerConfiguration));
+			Host.Start();
+			Configuration = conf;
+			_Client = NBXplorerNetwork.CreateExplorerClient(Address);
+			_Client.SetCookieAuth(Path.Combine(conf.DataDir, ".cookie"));
+			Notifications = _Client.CreateLongPollingNotificationSession();
+		}
+
+		string datadir;
+
+		public void ResetExplorer()
+		{
+			Host.Dispose();
+			DeleteFolderRecursive(datadir);
+			StartNBXplorer();
+		}
+
+		public LongPollingNotificationSession Notifications { get; set; }
 		private NetworkCredential ExtractCredentials(string config)
 		{
 			var user = Regex.Match(config, "rpcuser=([^\r\n]*)");
@@ -159,6 +187,13 @@ namespace NBXplorer.Tests
 			}
 		}
 
+		public T GetService<T>()
+		{
+			return ((T)(Host.Services.GetService(typeof(T))));
+		}
+
+		public ExplorerConfiguration Configuration { get; private set; }
+
 		ExplorerClient _Client;
 		public ExplorerClient Client
 		{
@@ -169,16 +204,6 @@ namespace NBXplorer.Tests
 		}
 
 		public CoreNode Explorer
-		{
-			get; set;
-		}
-
-		public CoreNode User1
-		{
-			get; set;
-		}
-
-		public CoreNode User2
 		{
 			get; set;
 		}
@@ -209,69 +234,61 @@ namespace NBXplorer.Tests
 
 		public Network Network
 		{
+			get
+			{
+				return NBXplorerNetwork.NBitcoinNetwork;
+			}
+		}
+		public NBXplorerNetwork NBXplorerNetwork
+		{
 			get;
 			internal set;
 		}
-		public bool RPCSupportSignRawTransaction { get; private set; } = true;
 
 		private static bool TryDelete(string directory, bool throws)
 		{
 			try
 			{
-				DeleteRecursivelyWithMagicDust(directory);
+				DeleteFolderRecursive(directory);
 				return true;
 			}
-			catch(DirectoryNotFoundException)
+			catch (DirectoryNotFoundException)
 			{
 				return true;
 			}
-			catch(Exception)
+			catch (Exception)
 			{
-				if(throws)
+				if (throws)
 					throw;
 			}
 			return false;
 		}
 
-		// http://stackoverflow.com/a/14933880/2061103
-		public static void DeleteRecursivelyWithMagicDust(string destinationDir)
+		public static void DeleteFolderRecursive(string destinationDir)
 		{
-			const int magicDust = 10;
-			for(var gnomes = 1; gnomes <= magicDust; gnomes++)
+			for (var i = 1; i <= 10; i++)
 			{
 				try
 				{
 					Directory.Delete(destinationDir, true);
+					return;
 				}
-				catch(DirectoryNotFoundException)
+				catch (DirectoryNotFoundException)
 				{
-					return;  // good!
+					return;
 				}
-				catch(IOException)
+				catch (IOException)
 				{
-					if(gnomes == magicDust)
-						throw;
-					// System.IO.IOException: The directory is not empty
-					System.Diagnostics.Debug.WriteLine("Gnomes prevent deletion of {0}! Applying magic dust, attempt #{1}.", destinationDir, gnomes);
-
-					// see http://stackoverflow.com/questions/329355/cannot-delete-directory-with-directory-deletepath-true for more magic
-					Thread.Sleep(100 * gnomes);
+					Thread.Sleep(100 * i);
 					continue;
 				}
-				catch(UnauthorizedAccessException)
+				catch (UnauthorizedAccessException)
 				{
-					if(gnomes == magicDust)
-						throw;
-					// Wait, maybe another software make us authorized a little later
-					System.Diagnostics.Debug.WriteLine("Gnomes prevent deletion of {0}! Applying magic dust, attempt #{1}.", destinationDir, gnomes);
-
-					// see http://stackoverflow.com/questions/329355/cannot-delete-directory-with-directory-deletepath-true for more magic
-					Thread.Sleep(100);
+					Thread.Sleep(100 * i);
 					continue;
 				}
-				return;
 			}
-			// depending on your use case, consider throwing an exception here
+			throw new IOException($"Impossible to delete folder {destinationDir}");
 		}
 
 		static void Copy(string sourceDirectory, string targetDirectory)
@@ -287,13 +304,13 @@ namespace NBXplorer.Tests
 			Directory.CreateDirectory(target.FullName);
 
 			// Copy each file into the new directory.
-			foreach(FileInfo fi in source.GetFiles())
+			foreach (FileInfo fi in source.GetFiles())
 			{
 				fi.CopyTo(Path.Combine(target.FullName, fi.Name), true);
 			}
 
 			// Copy each subdirectory using recursion.
-			foreach(DirectoryInfo diSourceSubDir in source.GetDirectories())
+			foreach (DirectoryInfo diSourceSubDir in source.GetDirectories())
 			{
 				DirectoryInfo nextTargetSubDir =
 					target.CreateSubdirectory(diSourceSubDir.Name);
@@ -308,7 +325,7 @@ namespace NBXplorer.Tests
 
 		public BitcoinAddress AddressOf(BitcoinExtKey key, string path)
 		{
-			if(SupportSegwit())
+			if (this.RPC.Capabilities.SupportSegwit)
 				return key.ExtKey.Derive(new KeyPath(path)).Neuter().PubKey.WitHash.GetAddress(Network);
 			else
 				return key.ExtKey.Derive(new KeyPath(path)).Neuter().PubKey.Hash.GetAddress(Network);
@@ -316,7 +333,7 @@ namespace NBXplorer.Tests
 
 		public BitcoinAddress AddressOf(DerivationStrategyBase scheme, string path)
 		{
-			return scheme.Derive(KeyPath.Parse(path)).ScriptPubKey.GetDestinationAddress(Network);
+			return scheme.GetDerivation(KeyPath.Parse(path)).ScriptPubKey.GetDestinationAddress(Network);
 		}
 
 		public DirectDerivationStrategy CreateDerivationStrategy(ExtPubKey pubKey = null)
@@ -326,41 +343,57 @@ namespace NBXplorer.Tests
 		public DerivationStrategyBase CreateDerivationStrategy(ExtPubKey pubKey, bool p2sh)
 		{
 			pubKey = pubKey ?? new ExtKey().Neuter();
-			string suffix = SupportSegwit() ? "" : "-[legacy]";
+			string suffix = this.RPC.Capabilities.SupportSegwit ? "" : "-[legacy]";
 			suffix += p2sh ? "-[p2sh]" : "";
-			return new DerivationStrategyFactory(this.Network).Parse($"{pubKey.ToString(this.Network)}{suffix}");
+			return NBXplorerNetwork.DerivationStrategyFactory.Parse($"{pubKey.ToString(this.Network)}{suffix}");
 		}
-
-		public bool RPCSupportSegwit
-		{
-			get; set;
-		} = true;
 
 		public bool RPCStringAmount
 		{
 			get; set;
 		} = true;
 
-		public bool SupportSegwit()
-		{
-			return RPCSupportSegwit && Network.Consensus.SupportSegwit;
-		}
-
 		public uint256 SendToAddress(BitcoinAddress address, Money amount)
 		{
 			return SendToAddressAsync(address, amount).GetAwaiter().GetResult();
+		}
+
+		public uint256 SendToAddress(Script scriptPubKey, Money amount)
+		{
+			return SendToAddressAsync(scriptPubKey.GetDestinationAddress(Network), amount).GetAwaiter().GetResult();
+		}
+		public Task<uint256> SendToAddressAsync(Script scriptPubKey, Money amount)
+		{
+			return SendToAddressAsync(scriptPubKey.GetDestinationAddress(Network), amount);
 		}
 
 		public async Task<uint256> SendToAddressAsync(BitcoinAddress address, Money amount)
 		{
 			List<object> parameters = new List<object>();
 			parameters.Add(address.ToString());
-			if(RPCStringAmount)
+			if (RPCStringAmount)
 				parameters.Add(amount.ToString());
 			else
 				parameters.Add(amount.ToDecimal(MoneyUnit.BTC));
 			var resp = await RPC.SendCommandAsync(RPCOperations.sendtoaddress, parameters.ToArray());
 			return uint256.Parse(resp.Result.ToString());
+		}
+
+		internal void WaitSynchronized()
+		{
+			using (var cts = new CancellationTokenSource(TimeSpan.FromSeconds(10)))
+			{
+				while (true)
+				{
+					cts.Token.ThrowIfCancellationRequested();
+					var status = Client.GetStatus();
+					if (status.SyncHeight == status.BitcoinStatus.Blocks)
+					{
+						break;
+					}
+					Thread.Sleep(50);
+				}
+			}
 		}
 	}
 }
