@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using NBXplorer.Models;
 
 namespace NBXplorer
 {
@@ -16,109 +17,209 @@ namespace NBXplorer
 	}
 	public class AnnotatedTransaction
 	{
-		public AnnotatedTransaction()
+		public AnnotatedTransaction(int? height, TrackedTransaction record, bool isMature)
 		{
-
-		}
-		public AnnotatedTransaction(TrackedTransaction tracked, SlimChain chain)
-		{
-			Record = tracked;
-			if(tracked.BlockHash == null)
-			{
-				Type = AnnotatedTransactionType.Unconfirmed;
-			}
-			else
-			{
-				var block = chain.GetBlock(tracked.BlockHash);
-				Type = block == null ? AnnotatedTransactionType.Orphan : AnnotatedTransactionType.Confirmed;
-				Height = block?.Height;
-			}
-		}
-		public AnnotatedTransactionType Type
-		{
-			get; set;
+			this.Record = record;
+			this.Height = height;
+			this.IsMature = isMature;
 		}
 		public int? Height
 		{
 			get;
-			internal set;
 		}
+		public bool IsMature { get; }
 		public TrackedTransaction Record
 		{
 			get;
-			internal set;
 		}
 
 		public override string ToString()
 		{
-			return Record?.Transaction?.GetHash()?.ToString() ?? "";
+			return Record.TransactionHash.ToString();
 		}
 	}
 
 	public class AnnotatedTransactionCollection : List<AnnotatedTransaction>
 	{
-		public AnnotatedTransactionCollection(IEnumerable<AnnotatedTransaction> transactions) : base(transactions)
+		public AnnotatedTransactionCollection(ICollection<TrackedTransaction> transactions, Models.TrackedSource trackedSource, SlimChain headerChain, Network network) : base(transactions.Count)
 		{
-			foreach(var tx in transactions)
+			_TxById = new Dictionary<uint256, AnnotatedTransaction>(transactions.Count);
+			foreach (var tx in transactions)
 			{
-				var h = tx.Record.Transaction.GetHash();
-				_TxById.Add(h, tx);
-				foreach(var keyPathInfo in tx.Record.TransactionMatch.Inputs.Concat(tx.Record.TransactionMatch.Outputs))
+				foreach (var keyPathInfo in tx.KnownKeyPathMapping)
 				{
-					if(keyPathInfo.KeyPath != null)
-						_KeyPaths.TryAdd(keyPathInfo.ScriptPubKey, keyPathInfo.KeyPath);
+					_KeyPaths.TryAdd(keyPathInfo.Key, keyPathInfo.Value);
 				}
 			}
 
-
-			UTXOState state = new UTXOState();
-			foreach(var confirmed in transactions
-										.Where(tx => tx.Type == AnnotatedTransactionType.Confirmed)
-										.TopologicalSort())
+			// Let's remove the dups and let's get the current height of the transactions
+			foreach (var trackedTx in transactions)
 			{
-				if(state.Apply(confirmed.Record.Transaction) == ApplyTransactionResult.Conflict
-					|| !ConfirmedTransactions.TryAdd(confirmed.Record.Transaction.GetHash(), confirmed))
-				{
-					Logs.Explorer.LogError("A conflict among confirmed transaction happened, this should be impossible");
-					throw new InvalidOperationException("The impossible happened");
-				}
-			}
+				int? txHeight = null;
+				bool isMature = true;
 
-			foreach(var unconfirmed in transactions
-										.Where(tx => tx.Type == AnnotatedTransactionType.Unconfirmed || tx.Type == AnnotatedTransactionType.Orphan)
-										.OrderByDescending(t => t.Record.Inserted) // OrderByDescending so that the last received is least likely to be conflicted
-										.TopologicalSort())
-			{
-				var hash = unconfirmed.Record.Transaction.GetHash();
-				if(ConfirmedTransactions.ContainsKey(hash))
+				if (trackedTx.BlockHash != null && headerChain.TryGetHeight(trackedTx.BlockHash, out var height))
 				{
-					DuplicatedTransactions.Add(unconfirmed);
+					txHeight = height;
+					isMature = trackedTx.IsCoinBase ? headerChain.Height - height >= network.Consensus.CoinbaseMaturity : true;
 				}
-				else
+
+				AnnotatedTransaction annotatedTransaction = new AnnotatedTransaction(txHeight, trackedTx, isMature);
+				if (_TxById.TryGetValue(trackedTx.TransactionHash, out var conflicted))
 				{
-					if(state.Apply(unconfirmed.Record.Transaction) == ApplyTransactionResult.Conflict)
+					if (ShouldReplace(annotatedTransaction, conflicted))
 					{
-						ReplacedTransactions.TryAdd(hash, unconfirmed);
+						CleanupTransactions.Add(conflicted);
+						_TxById.Remove(trackedTx.TransactionHash);
+						_TxById.Add(trackedTx.TransactionHash, annotatedTransaction);
 					}
 					else
 					{
-						if(!UnconfirmedTransactions.TryAdd(hash, unconfirmed))
+						CleanupTransactions.Add(annotatedTransaction);
+					}
+				}
+				else
+				{
+					_TxById.Add(trackedTx.TransactionHash, annotatedTransaction);
+				}
+			}
+
+			// Let's resolve the double spents
+			Dictionary<OutPoint, uint256> spentBy = new Dictionary<OutPoint, uint256>(transactions.SelectMany(t => t.SpentOutpoints).Count());
+			foreach (var annotatedTransaction in _TxById.Values.Where(r => r.Height is int))
+			{
+				foreach (var spent in annotatedTransaction.Record.SpentOutpoints)
+				{
+					// No way to have double spent in confirmed transactions
+					spentBy.Add(spent, annotatedTransaction.Record.TransactionHash);
+				}
+			}
+
+		removeConflicts:
+			HashSet<uint256> toRemove = new HashSet<uint256>();
+			foreach (var annotatedTransaction in _TxById.Values.Where(r => r.Height is null))
+			{
+				foreach (var spent in annotatedTransaction.Record.SpentOutpoints)
+				{
+					if (spentBy.TryGetValue(spent, out var conflictHash) &&
+						_TxById.TryGetValue(conflictHash, out var conflicted))
+					{
+						if (conflicted == annotatedTransaction)
+							goto nextTransaction;
+						if (toRemove.Contains(conflictHash))
 						{
-							throw new InvalidOperationException("The impossible happened (!UnconfirmedTransactions.TryAdd(hash, unconfirmed))");
+							spentBy.Remove(spent);
+							spentBy.Add(spent, annotatedTransaction.Record.TransactionHash);
+						}
+						else if (ShouldReplace(annotatedTransaction, conflicted))
+						{
+							toRemove.Add(conflictHash);
+							spentBy.Remove(spent);
+							spentBy.Add(spent, annotatedTransaction.Record.TransactionHash);
+
+							if (conflicted.Height is null && annotatedTransaction.Height is null)
+							{
+								ReplacedTransactions.Add(conflicted);
+							}
+							else
+							{
+								CleanupTransactions.Add(conflicted);
+							}
+						}
+						else
+						{
+							toRemove.Add(annotatedTransaction.Record.TransactionHash);
+							if (conflicted.Height is null && annotatedTransaction.Height is null)
+							{
+								ReplacedTransactions.Add(annotatedTransaction);
+							}
+							else
+							{
+								CleanupTransactions.Add(annotatedTransaction);
+							}
 						}
 					}
+					else
+					{
+						spentBy.Add(spent, annotatedTransaction.Record.TransactionHash);
+					}
+				}
+			nextTransaction:;
+			}
+
+			foreach (var e in toRemove)
+				_TxById.Remove(e);
+			if (toRemove.Count != 0)
+				goto removeConflicts;
+
+			var sortedTransactions = _TxById.Values.TopologicalSort();
+			UTXOState state = new UTXOState();
+			foreach (var tx in sortedTransactions.Where(s => s.IsMature && s.Height is int))
+			{
+				if (state.Apply(tx.Record) == ApplyTransactionResult.Conflict)
+				{
+					throw new InvalidOperationException("The impossible happened");
+				}
+			}
+			ConfirmedState = state.Snapshot();
+			foreach (var tx in sortedTransactions.Where(s => s.Height is null))
+			{
+				if (state.Apply(tx.Record) == ApplyTransactionResult.Conflict)
+				{
+					throw new InvalidOperationException("The impossible happened");
+				}
+			}
+			foreach (var tx in sortedTransactions)
+			{
+				if (tx.Height is int)
+					ConfirmedTransactions.Add(tx);
+				else
+					UnconfirmedTransactions.Add(tx);
+				this.Add(tx);
+			}
+			UnconfirmedState = state;
+			TrackedSource = trackedSource;
+		}
+
+		public UTXOState ConfirmedState { get; set; }
+		public UTXOState UnconfirmedState { get; set; }
+
+		private static bool ShouldReplace(AnnotatedTransaction annotatedTransaction, AnnotatedTransaction conflicted)
+		{
+			if (annotatedTransaction.Height is null)
+			{
+				if (conflicted.Height is null)
+				{
+					return annotatedTransaction.Record.FirstSeen > conflicted.Record.FirstSeen; // The is a replaced tx, we want the youngest
+				}
+				else
+				{
+					return false;
+				}
+			}
+			else
+			{
+				if (conflicted.Height is null)
+				{
+					return true;
+				}
+				else
+				{
+					if (annotatedTransaction.Record.Key.TxId == conflicted.Record.Key.TxId && 
+						annotatedTransaction.Height == conflicted.Height)
+					{
+						return !annotatedTransaction.Record.Key.IsPruned;
+					}
+					return annotatedTransaction.Height < conflicted.Height; // The most buried block win (should never happen though)
 				}
 			}
 		}
 
-		public TxOut GetUTXO(OutPoint outpoint)
+		public MatchedOutput GetUTXO(OutPoint outpoint)
 		{
-			if(_TxById.TryGetValue(outpoint.Hash, out var txs))
+			if (_TxById.TryGetValue(outpoint.Hash, out var tx))
 			{
-				var tx = txs.First().Record.Transaction;
-				if(outpoint.N >= tx.Outputs.Count)
-					return null;
-				return tx.Outputs[outpoint.N];
+				return tx.Record.GetReceivedOutputs().Where(c => c.Index == outpoint.N).FirstOrDefault();
 			}
 			return null;
 		}
@@ -129,36 +230,35 @@ namespace NBXplorer
 			return _KeyPaths.TryGet(scriptPubkey);
 		}
 
-		MultiValueDictionary<uint256, AnnotatedTransaction> _TxById = new MultiValueDictionary<uint256, AnnotatedTransaction>();
-		public IReadOnlyCollection<AnnotatedTransaction> GetByTxId(uint256 txId)
+		Dictionary<uint256, AnnotatedTransaction> _TxById = new Dictionary<uint256, AnnotatedTransaction>();
+		public AnnotatedTransaction GetByTxId(uint256 txId)
 		{
-			if(txId == null)
+			if (txId == null)
 				throw new ArgumentNullException(nameof(txId));
-			IReadOnlyCollection<AnnotatedTransaction> value;
-			if(_TxById.TryGetValue(txId, out value))
-				return value;
-			return new List<AnnotatedTransaction>();
+			_TxById.TryGetValue(txId, out var value);
+			return value;
 		}
 
-		public Dictionary<uint256, AnnotatedTransaction> ReplacedTransactions
-		{
-			get; set;
-		} = new Dictionary<uint256, AnnotatedTransaction>();
-
-		public Dictionary<uint256, AnnotatedTransaction> ConfirmedTransactions
-		{
-			get; set;
-		} = new Dictionary<uint256, AnnotatedTransaction>();
-
-		public Dictionary<uint256, AnnotatedTransaction> UnconfirmedTransactions
-		{
-			get; set;
-		} = new Dictionary<uint256, AnnotatedTransaction>();
-
-		public List<AnnotatedTransaction> DuplicatedTransactions
+		public List<AnnotatedTransaction> ReplacedTransactions
 		{
 			get; set;
 		} = new List<AnnotatedTransaction>();
+
+		public List<AnnotatedTransaction> ConfirmedTransactions
+		{
+			get; set;
+		} = new List<AnnotatedTransaction>();
+
+		public List<AnnotatedTransaction> UnconfirmedTransactions
+		{
+			get; set;
+		} = new List<AnnotatedTransaction>();
+
+		public List<AnnotatedTransaction> CleanupTransactions
+		{
+			get; set;
+		} = new List<AnnotatedTransaction>();
+		public Models.TrackedSource TrackedSource { get; }
 	}
 
 }
