@@ -15,6 +15,7 @@ using NBitcoin.Protocol.Behaviors;
 using Microsoft.Extensions.Hosting;
 using NBXplorer.Events;
 using Newtonsoft.Json.Linq;
+using System.Text;
 
 namespace NBXplorer
 {
@@ -26,22 +27,12 @@ namespace NBXplorer
 		Ready
 	}
 
-	/// <summary>
-	/// Hack, ASP.NET core DI does not support having one singleton for multiple interfaces
-	/// </summary>
-	public class BitcoinDWaitersAccessor
-	{
-		public BitcoinDWaiters Instance
-		{
-			get; set;
-		}
-	}
-
 	public class BitcoinDWaiters : IHostedService
 	{
 		Dictionary<string, BitcoinDWaiter> _Waiters;
+		private readonly RepositoryProvider repositoryProvider;
+
 		public BitcoinDWaiters(
-							BitcoinDWaitersAccessor accessor,
 							AddressPoolServiceAccessor addressPool,
 								NBXplorerNetworkProvider networkProvider,
 							  ChainProvider chains,
@@ -50,7 +41,6 @@ namespace NBXplorer
 							  RPCClientProvider rpcProvider,
 							  EventAggregator eventAggregator)
 		{
-			accessor.Instance = this;
 			_Waiters = networkProvider
 				.GetAll()
 				.Select(s => (Repository: repositoryProvider.GetRepository(s),
@@ -66,15 +56,18 @@ namespace NBXplorer
 												addressPool.Instance,
 												eventAggregator))
 				.ToDictionary(s => s.Network.CryptoCode, s => s);
+			this.repositoryProvider = repositoryProvider;
 		}
-		public Task StartAsync(CancellationToken cancellationToken)
+		public async Task StartAsync(CancellationToken cancellationToken)
 		{
-			return Task.WhenAll(_Waiters.Select(s => s.Value.StartAsync(cancellationToken)).ToArray());
+			await repositoryProvider.StartAsync();
+			await Task.WhenAll(_Waiters.Select(s => s.Value.StartAsync(cancellationToken)).ToArray());
 		}
 
-		public Task StopAsync(CancellationToken cancellationToken)
+		public async Task StopAsync(CancellationToken cancellationToken)
 		{
-			return Task.WhenAll(_Waiters.Select(s => s.Value.StopAsync(cancellationToken)).ToArray());
+			await Task.WhenAll(_Waiters.Select(s => s.Value.StopAsync(cancellationToken)).ToArray());
+			await repositoryProvider.DisposeAsync();
 		}
 
 		public BitcoinDWaiter GetWaiter(NBXplorerNetwork network)
@@ -83,7 +76,7 @@ namespace NBXplorer
 		}
 		public BitcoinDWaiter GetWaiter(string cryptoCode)
 		{
-			_Waiters.TryGetValue(cryptoCode, out BitcoinDWaiter waiter);
+			_Waiters.TryGetValue(cryptoCode.ToUpperInvariant(), out BitcoinDWaiter waiter);
 			return waiter;
 		}
 
@@ -95,14 +88,15 @@ namespace NBXplorer
 
 	public class BitcoinDWaiter : IHostedService
 	{
-		RPCClient _RPC;
+		RPCClient _RPCWithTimeout;
+		RPCClient _OriginalRPC;
 		NBXplorerNetwork _Network;
 		ExplorerConfiguration _Configuration;
-		private readonly AddressPoolService _AddressPoolService;
+		private ExplorerBehavior _ExplorerPrototype;
 		SlimChain _Chain;
-		private Repository _Repository;
 		EventAggregator _EventAggregator;
 		private readonly ChainConfiguration _ChainConfiguration;
+		readonly string RPCReadyFile;
 
 		public BitcoinDWaiter(
 			RPCClient rpc,
@@ -113,17 +107,20 @@ namespace NBXplorer
 			AddressPoolService addressPoolService,
 			EventAggregator eventAggregator)
 		{
-			if(addressPoolService == null)
+			if (addressPoolService == null)
 				throw new ArgumentNullException(nameof(addressPoolService));
-			_RPC = rpc;
+			_OriginalRPC = rpc;
+			_RPCWithTimeout = rpc.Clone();
+			_RPCWithTimeout.RequestTimeout = TimeSpan.FromMinutes(1.0);
 			_Configuration = configuration;
-			_AddressPoolService = addressPoolService;
 			_Network = network;
 			_Chain = chain;
-			_Repository = repository;
 			State = BitcoinDWaiterState.NotStarted;
 			_EventAggregator = eventAggregator;
 			_ChainConfiguration = _Configuration.ChainConfigurations.First(c => c.CryptoCode == _Network.CryptoCode);
+			_ExplorerPrototype = new ExplorerBehavior(repository, chain, addressPoolService, eventAggregator) { StartHeight = _ChainConfiguration.StartHeight };
+			RPCReadyFile = Path.Combine(configuration.SignalFilesDir, $"{network.CryptoCode.ToLowerInvariant()}_fully_synched");
+			HasTxIndex = _ChainConfiguration.HasTxIndex;
 		}
 		public NodeState NodeState
 		{
@@ -131,7 +128,7 @@ namespace NBXplorer
 			private set;
 		}
 
-		private NodesGroup _Group;
+		private Node _Node;
 
 
 		public NBXplorerNetwork Network
@@ -146,7 +143,7 @@ namespace NBXplorer
 		{
 			get
 			{
-				return _RPC;
+				return _RPCWithTimeout;
 			}
 		}
 
@@ -172,48 +169,65 @@ namespace NBXplorer
 		CancellationTokenSource _Cts;
 		public Task StartAsync(CancellationToken cancellationToken)
 		{
-			if(_Disposed)
+			if (_Disposed)
 				throw new ObjectDisposedException(nameof(BitcoinDWaiter));
 
 			_Cts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
 			_Loop = StartLoop(_Cts.Token, _Tick);
-			_Subscription = _EventAggregator.Subscribe<NewBlockEvent>(s =>
+			_Subscription = _EventAggregator.Subscribe<Models.NewBlockEvent>(s =>
 			{
 				_Tick.Set();
 			});
 			return Task.CompletedTask;
 		}
 
-		AutoResetEvent _Tick = new AutoResetEvent(false);
+		Signaler _Tick = new Signaler();
 
-		private async Task StartLoop(CancellationToken token, AutoResetEvent tick)
+		private async Task StartLoop(CancellationToken token, Signaler tick)
 		{
 			try
 			{
-				while(!token.IsCancellationRequested)
+				int errors = 0;
+				while (!token.IsCancellationRequested)
 				{
+					errors = Math.Min(11, errors);
 					try
 					{
-						while(await StepAsync(token))
+						while (await StepAsync(token))
 						{
 						}
-						await Task.WhenAny(tick.WaitOneAsync(), Task.Delay(PollingInterval, token));
+						await tick.Wait(PollingInterval, token);
+						errors = 0;
 					}
-					catch(Exception ex) when(!token.IsCancellationRequested)
+					catch (ConfigException) when (!token.IsCancellationRequested)
+					{
+						// Probably RPC errors, don't spam
+						await Wait(errors, tick, token);
+						errors++;
+					}
+					catch (Exception ex) when (!token.IsCancellationRequested)
 					{
 						Logs.Configuration.LogError(ex, $"{_Network.CryptoCode}: Unhandled in Waiter loop");
-						await Task.WhenAny(tick.WaitOneAsync(), Task.Delay(TimeSpan.FromSeconds(5.0), token));
+						await Wait(errors, tick, token);
+						errors++;
 					}
 				}
 			}
-			catch when(token.IsCancellationRequested)
+			catch when (token.IsCancellationRequested)
 			{
 			}
 		}
 
+		private async Task Wait(int errors, Signaler tick, CancellationToken token)
+		{
+			var timeToWait = TimeSpan.FromSeconds(5.0) * (errors + 1);
+			Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Testing again in {(int)timeToWait.TotalSeconds} seconds");
+			await tick.Wait(timeToWait, token);
+		}
+
 		public BlockLocator GetLocation()
 		{
-			return _Group?.ConnectedNodes?.FirstOrDefault()?.Behaviors?.Find<ExplorerBehavior>()?.CurrentLocation;
+			return GetExplorerBehavior()?.CurrentLocation;
 		}
 
 		public TimeSpan PollingInterval
@@ -221,60 +235,55 @@ namespace NBXplorer
 			get; set;
 		} = TimeSpan.FromMinutes(1.0);
 
-		private void ConnectedNodes_Changed(object sender, NodeEventArgs e)
-		{
-			_Tick.Set();
-		}
-
-		public Task StopAsync(CancellationToken cancellationToken)
+		public async Task StopAsync(CancellationToken cancellationToken)
 		{
 			_Disposed = true;
 			_Cts.Cancel();
 			_Subscription.Dispose();
-			if(_Group != null)
-			{
-				_Group.ConnectedNodes.Added -= ConnectedNodes_Changed;
-				_Group.ConnectedNodes.Removed -= ConnectedNodes_Changed;
-				_Group.Disconnect();
-				_Group = null;
-			}
+			EnsureNodeDisposed();
 			State = BitcoinDWaiterState.NotStarted;
 			_Chain = null;
-			_Tick.Set();
-			_Tick.Dispose();
 			try
 			{
-				_Loop.Wait();
+				await _Loop;
 			}
 			catch { }
-			return _Loop;
+			EnsureRPCReadyFileDeleted();
 		}
-
+		bool _BanListLoaded;
 		async Task<bool> StepAsync(CancellationToken token)
 		{
 			var oldState = State;
-			switch(State)
+			switch (State)
 			{
 				case BitcoinDWaiterState.NotStarted:
-					await RPCArgs.TestRPCAsync(_Network, _RPC, token);
+					await RPCArgs.TestRPCAsync(_Network, _RPCWithTimeout, token);
+					_OriginalRPC.Capabilities = _RPCWithTimeout.Capabilities;
 					GetBlockchainInfoResponse blockchainInfo = null;
 					try
 					{
-						blockchainInfo = await _RPC.GetBlockchainInfoAsyncEx();
-						if(blockchainInfo != null && _Network.NBitcoinNetwork.NetworkType == NetworkType.Regtest)
+						blockchainInfo = await _RPCWithTimeout.GetBlockchainInfoAsyncEx();
+						if (_Network.CryptoCode == "BTC" &&
+							_Network.NBitcoinNetwork.NetworkType == NetworkType.Mainnet &&
+							!_BanListLoaded)
 						{
-							if(await WarmupBlockchain())
+							if (await LoadBanList())
+								_BanListLoaded = true;
+						}
+						if (blockchainInfo != null && _Network.NBitcoinNetwork.NetworkType == NetworkType.Regtest)
+						{
+							if (await WarmupBlockchain())
 							{
-								blockchainInfo = await _RPC.GetBlockchainInfoAsyncEx();
+								blockchainInfo = await _RPCWithTimeout.GetBlockchainInfoAsyncEx();
 							}
 						}
 					}
-					catch(Exception ex)
+					catch (Exception ex)
 					{
 						Logs.Configuration.LogError(ex, $"{_Network.CryptoCode}: Failed to connect to RPC");
 						break;
 					}
-					if(IsSynchingCore(blockchainInfo))
+					if (IsSynchingCore(blockchainInfo))
 					{
 						State = BitcoinDWaiterState.CoreSynching;
 					}
@@ -288,50 +297,38 @@ namespace NBXplorer
 					GetBlockchainInfoResponse blockchainInfo2 = null;
 					try
 					{
-						blockchainInfo2 = await _RPC.GetBlockchainInfoAsyncEx();
+						blockchainInfo2 = await _RPCWithTimeout.GetBlockchainInfoAsyncEx();
 					}
-					catch(Exception ex)
+					catch (Exception ex)
 					{
 						Logs.Configuration.LogError(ex, $"{_Network.CryptoCode}: Failed to connect to RPC");
 						State = BitcoinDWaiterState.NotStarted;
 						break;
 					}
-					if(!IsSynchingCore(blockchainInfo2))
+					if (!IsSynchingCore(blockchainInfo2))
 					{
 						await ConnectToBitcoinD(token);
 						State = BitcoinDWaiterState.NBXplorerSynching;
 					}
 					break;
 				case BitcoinDWaiterState.NBXplorerSynching:
-					var explorer = _Group?.ConnectedNodes.SelectMany(n => n.Behaviors.OfType<ExplorerBehavior>()).FirstOrDefault();
-					if(explorer == null)
+					var explorer = GetExplorerBehavior();
+					if (explorer == null)
 					{
-						GetBlockchainInfoResponse blockchainInfo3 = null;
-						try
-						{
-							blockchainInfo3 = await _RPC.GetBlockchainInfoAsyncEx();
-						}
-						catch(Exception ex)
-						{
-							Logs.Configuration.LogError(ex, $"{_Network.CryptoCode}: Failed to connect to RPC");
-							State = BitcoinDWaiterState.NotStarted;
-							break;
-						}
-						if(IsSynchingCore(blockchainInfo3))
-							State = BitcoinDWaiterState.CoreSynching;
+						State = BitcoinDWaiterState.NotStarted;
 					}
-					else if(!explorer.IsSynching())
+					else if (!explorer.IsSynching())
 					{
 						State = BitcoinDWaiterState.Ready;
 					}
 					break;
 				case BitcoinDWaiterState.Ready:
-					var explorer2 = _Group?.ConnectedNodes.SelectMany(n => n.Behaviors.OfType<ExplorerBehavior>()).FirstOrDefault();
-					if(explorer2 == null)
+					var explorer2 = GetExplorerBehavior();
+					if (explorer2 == null)
 					{
 						State = BitcoinDWaiterState.NotStarted;
 					}
-					else if(explorer2.IsSynching())
+					else if (explorer2.IsSynching())
 					{
 						State = BitcoinDWaiterState.NBXplorerSynching;
 					}
@@ -341,47 +338,212 @@ namespace NBXplorer
 			}
 			var changed = oldState != State;
 
-			if(changed)
+			if (changed)
 			{
+				if (oldState == BitcoinDWaiterState.NotStarted)
+					NetworkInfo = await _RPCWithTimeout.GetNetworkInfoAsync();
 				_EventAggregator.Publish(new BitcoinDStateChangedEvent(_Network, oldState, State));
+				if (State == BitcoinDWaiterState.Ready)
+				{
+					await File.WriteAllTextAsync(RPCReadyFile, NBitcoin.Utils.DateTimeToUnixTime(DateTimeOffset.UtcNow).ToString());
+				}
 			}
-
+			if (State != BitcoinDWaiterState.Ready)
+			{
+				EnsureRPCReadyFileDeleted();
+			}
 			return changed;
+		}
+
+		private Node GetHandshakedNode()
+		{
+			return _Node?.State == NodeState.HandShaked ? _Node : null;
+		}
+
+		private ExplorerBehavior GetExplorerBehavior()
+		{
+			return GetHandshakedNode()?.Behaviors?.Find<ExplorerBehavior>();
+		}
+
+		private void EnsureRPCReadyFileDeleted()
+		{
+			if (File.Exists(RPCReadyFile))
+				File.Delete(RPCReadyFile);
+		}
+
+		private async Task<bool> LoadBanList()
+		{
+			var stream = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceStream("NBXplorer.banlist.cli.txt");
+			string content = null;
+			using (var reader = new StreamReader(stream, Encoding.UTF8))
+			{
+				content = reader.ReadToEnd();
+			}
+			var bannedLines = content.Split(new[] { "\r\n", "\n" }, StringSplitOptions.RemoveEmptyEntries);
+			var batch = _RPCWithTimeout.PrepareBatch();
+			var commands = bannedLines
+						.Where(o => o.Length > 0 && o[0] != '#')
+						.Select(b => b.Split(' ')[2])
+						.Select(ip => batch.SendCommandAsync(new RPCRequest("setban", new object[] { ip, "add", 31557600 }), false))
+						.ToArray();
+			await batch.SendBatchAsync();
+			foreach (var command in commands)
+			{
+				var result = await command;
+				if (result.Error != null && result.Error.Code != RPCErrorCode.RPC_CLIENT_NODE_ALREADY_ADDED)
+					result.ThrowIfError();
+			}
+			Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Node banlist loaded");
+			return true;
 		}
 
 		private async Task ConnectToBitcoinD(CancellationToken cancellation)
 		{
-			if(_Group != null)
+			var node = GetHandshakedNode();
+			if (node != null)
 				return;
-			_Chain.ResetToGenesis();
-			if (_Configuration.CacheChain)
+			try
 			{
-				LoadChainFromCache();
-				if (!await HasBlock(_RPC, _Chain.Tip))
+				EnsureNodeDisposed();
+				_Chain.ResetToGenesis();
+				if (_Configuration.CacheChain)
 				{
-					Logs.Configuration.LogInformation($"{_Network.CryptoCode}: The cached chain contains a tip unknown to the node, dropping the cache...");
-					_Chain.ResetToGenesis();
+					LoadChainFromCache();
+					if (!await HasBlock(_RPCWithTimeout, _Chain.Tip))
+					{
+						Logs.Configuration.LogInformation($"{_Network.CryptoCode}: The cached chain contains a tip unknown to the node, dropping the cache...");
+						_Chain.ResetToGenesis();
+					}
 				}
+				var heightBefore = _Chain.Height;
+				using (var timeout = CancellationTokenSource.CreateLinkedTokenSource(cancellation))
+				{
+					timeout.CancelAfter(_Network.ChainLoadingTimeout);
+					try
+					{
+						Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Trying to connect via the P2P protocol to trusted node ({_ChainConfiguration.NodeEndpoint.ToEndpointString()})...");
+						var userAgent = "NBXplorer-" + RandomUtils.GetInt64();
+						bool handshaked = false;
+						using (var handshakeTimeout = CancellationTokenSource.CreateLinkedTokenSource(cancellation))
+						{
+							try
+							{
+								handshakeTimeout.CancelAfter(TimeSpan.FromSeconds(10));
+								node = await Node.ConnectAsync(_Network.NBitcoinNetwork, _ChainConfiguration.NodeEndpoint, new NodeConnectionParameters()
+								{
+									UserAgent = userAgent,
+									ConnectCancellation = handshakeTimeout.Token,
+									IsRelay = true
+								});
+
+								try
+								{
+									Logs.Explorer.LogInformation($"{Network.CryptoCode}: TCP Connection succeed, handshaking...");
+									node.VersionHandshake(handshakeTimeout.Token);
+									Logs.Explorer.LogInformation($"{Network.CryptoCode}: Handshaked");
+								}
+								catch (OperationCanceledException) when (handshakeTimeout.IsCancellationRequested)
+								{
+									Logs.Explorer.LogWarning($"{Network.CryptoCode}: NBXplorer could not complete the handshake with the remote node. This is probably because NBXplorer is not whitelisted by your node.{Environment.NewLine}" +
+										$"You can use \"whitebind\" or \"whitelist\" in your node configuration. (typically whitelist=127.0.0.1 if NBXplorer and the node are on the same machine.){Environment.NewLine}" +
+										$"This issue can also happen because NBXplorer do not manage to connect to the P2P port of your node at all.");
+									throw;
+								}
+								handshaked = true;
+								var loadChainTimeout = _Network.NBitcoinNetwork.NetworkType == NetworkType.Regtest ? TimeSpan.FromSeconds(5) : _Network.ChainCacheLoadingTimeout;
+								if (_Chain.Height < 5)
+									loadChainTimeout = TimeSpan.FromDays(7); // unlimited
+								Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Loading chain from node");
+								try
+								{
+									using (var cts1 = CancellationTokenSource.CreateLinkedTokenSource(cancellation))
+									{
+										cts1.CancelAfter(loadChainTimeout);
+										Logs.Explorer.LogInformation($"{Network.CryptoCode}: Loading chain...");
+										node.SynchronizeSlimChain(_Chain, cancellationToken: cts1.Token);
+									}
+								}
+								catch when (!cancellation.IsCancellationRequested) // Timeout happens with SynchronizeChain, if so, throw away the cached chain
+								{
+									Logs.Explorer.LogInformation($"{Network.CryptoCode}: Failed to load chain before timeout, let's try again without the chain cache...");
+									_Chain.ResetToGenesis();
+									node.SynchronizeSlimChain(_Chain, cancellationToken: cancellation);
+								}
+								Logs.Explorer.LogInformation($"{Network.CryptoCode}: Chain loaded");
+
+								var peer = (await _RPCWithTimeout.GetPeersInfoAsync())
+											.FirstOrDefault(p => p.SubVersion == userAgent);
+								if (peer != null && !peer.IsWhiteListed)
+								{
+									var addressStr = peer.Address?.Address?.ToString();
+									if (addressStr == null)
+									{
+										addressStr = peer.AddressString;
+										var portDelimiter = addressStr.LastIndexOf(':');
+										if (portDelimiter != -1)
+											addressStr = addressStr.Substring(0, portDelimiter);
+									}
+
+									Logs.Explorer.LogWarning($"{Network.CryptoCode}: Your NBXplorer server is not whitelisted by your node," +
+										$" you should add \"whitelist={addressStr}\" to the configuration file of your node. (Or use whitebind)");
+								}
+								if (peer != null && peer.IsWhiteListed)
+								{
+									Logs.Explorer.LogInformation($"{Network.CryptoCode}: NBXplorer is correctly whitelisted by the node");
+								}
+							}
+							catch (OperationCanceledException) when (!handshaked && handshakeTimeout.IsCancellationRequested)
+							{
+								Logs.Explorer.LogWarning($"{Network.CryptoCode}: The initial handshake failed, your NBXplorer server might not be whitelisted by your node," +
+										$" if your bitcoin node is on the same machine as NBXplorer, you should add \"whitelist=127.0.0.1\" to the configuration file of your node. (Or use whitebind)");
+								throw;
+							}
+						}
+						Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Height: " + _Chain.Height);
+					}
+					catch (Exception ex) when (!cancellation.IsCancellationRequested)
+					{
+						throw new OperationCanceledException("Loading the chain from the node timed out", ex, timeout.Token);
+					}
+				}
+				if (_Configuration.CacheChain && heightBefore != _Chain.Height)
+				{
+					SaveChainInCache();
+				}
+				GC.Collect();
+				node.Behaviors.Add(new SlimChainBehavior(_Chain));
+				var explorer = (ExplorerBehavior)_ExplorerPrototype.Clone();
+				node.Behaviors.Add(explorer);
+				node.StateChanged += Node_StateChanged;
+				_Node = node;
+				await explorer.Init();
 			}
-			var heightBefore = _Chain.Height;
-			using(var timeout = CancellationTokenSource.CreateLinkedTokenSource(cancellation))
+			catch
 			{
-				timeout.CancelAfter(_Network.ChainLoadingTimeout);
+				EnsureNodeDisposed(node ?? _Node);
+				throw;
+			}
+		}
+
+		private void Node_StateChanged(Node node, NodeState oldState)
+		{
+			_Tick.Set();
+		}
+
+		private void EnsureNodeDisposed(Node node = null)
+		{
+			node = node ?? _Node;
+			if (node != null)
+			{
 				try
 				{
-					await LoadChainFromNode(timeout.Token);
+					node.StateChanged -= Node_StateChanged;
+					node.DisconnectAsync();
 				}
-				catch when(!cancellation.IsCancellationRequested)
-				{
-					throw new OperationCanceledException("Loading the chain from the node timed out", timeout.Token);
-				}
+				catch { }
+				node = null;
+				_Node = null;
 			}
-			if(_Configuration.CacheChain && heightBefore != _Chain.Height)
-			{
-				SaveChainInCache();
-			}
-			GC.Collect();
-			await LoadGroup();
 		}
 
 		private async Task<bool> HasBlock(RPCClient rpc, uint256 tip)
@@ -409,65 +571,6 @@ namespace NBXplorer
 			}
 		}
 
-		private async Task LoadGroup()
-		{
-			AddressManager manager = new AddressManager();
-			manager.Add(new NetworkAddress(_ChainConfiguration.NodeEndpoint), IPAddress.Loopback);
-			NodesGroup group = new NodesGroup(_Network.NBitcoinNetwork, new NodeConnectionParameters()
-			{
-				Services = NodeServices.Nothing,
-				IsRelay = true,
-				TemplateBehaviors =
-				{
-					new AddressManagerBehavior(manager)
-					{
-						PeersToDiscover = 1,
-						Mode = AddressManagerBehaviorMode.None
-					},
-					new ExplorerBehavior(_Repository, _Chain, _AddressPoolService, _EventAggregator) { StartHeight = _ChainConfiguration.StartHeight },
-					new SlimChainBehavior(_Chain),
-					new PingPongBehavior()
-				}
-			});
-			group.AllowSameGroup = true;
-			group.MaximumNodeConnection = 1;
-
-			var task = WaitConnected(group);
-
-			group.Connect();
-
-			try
-			{
-				await task;
-			}
-			catch(Exception ex)
-			{
-				Logs.Configuration.LogError(ex, $"{_Network.CryptoCode}: Failure to connect to the bitcoin node (P2P)");
-				throw;
-			}
-			_Group = group;
-
-			group.ConnectedNodes.Added += ConnectedNodes_Changed;
-			group.ConnectedNodes.Removed += ConnectedNodes_Changed;
-		}
-
-		private static async Task WaitConnected(NodesGroup group)
-		{
-			TaskCompletionSource<bool> tcs = new TaskCompletionSource<bool>();
-			EventHandler<NodeEventArgs> waitingConnected = null;
-			waitingConnected = (a, b) =>
-			{
-				tcs.TrySetResult(true);
-				group.ConnectedNodes.Added -= waitingConnected;
-			};
-			group.ConnectedNodes.Added += waitingConnected;
-			CancellationTokenSource cts = new CancellationTokenSource(5000);
-			using(cts.Token.Register(() => tcs.TrySetCanceled()))
-			{
-				await tcs.Task;
-			}
-		}
-
 		private void SaveChainInCache()
 		{
 			var suffix = _Network.CryptoCode == "BTC" ? "" : _Network.CryptoCode;
@@ -475,73 +578,16 @@ namespace NBXplorer
 			var cachePathTemp = Path.Combine(_Configuration.DataDir, $"{suffix}chain-slim.dat.temp");
 
 			Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Saving chain to cache...");
-			using(var fs = new FileStream(cachePathTemp, FileMode.Create, FileAccess.Write, FileShare.None, 1024 * 1024))
+			using (var fs = new FileStream(cachePathTemp, FileMode.Create, FileAccess.Write, FileShare.None, 1024 * 1024))
 			{
 				_Chain.Save(fs);
 				fs.Flush();
 			}
 
-			if(File.Exists(cachePath))
+			if (File.Exists(cachePath))
 				File.Delete(cachePath);
 			File.Move(cachePathTemp, cachePath);
 			Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Chain cached");
-		}
-
-		private async Task LoadChainFromNode(CancellationToken cancellation)
-		{
-			Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Loading chain from node ({_ChainConfiguration.NodeEndpoint.Address.ToString()}:{_ChainConfiguration.NodeEndpoint.Port})...");
-			var userAgent = "NBXplorer-" + RandomUtils.GetInt64();
-			bool handshaked = false;
-			using(var handshakeTimeout = CancellationTokenSource.CreateLinkedTokenSource(cancellation))
-			{
-				try
-				{
-					handshakeTimeout.CancelAfter(TimeSpan.FromSeconds(10));
-					using(var node = Node.Connect(_Network.NBitcoinNetwork, _ChainConfiguration.NodeEndpoint, new NodeConnectionParameters()
-					{
-						UserAgent = userAgent,
-						ConnectCancellation = handshakeTimeout.Token,
-						IsRelay = false
-					}))
-					{
-						node.VersionHandshake(handshakeTimeout.Token);
-						handshaked = true;
-						var loadChainTimeout = _Network.NBitcoinNetwork.NetworkType == NetworkType.Regtest ? TimeSpan.FromSeconds(5) : TimeSpan.FromSeconds(15);
-						if(_Chain.Height < 5)
-							loadChainTimeout = TimeSpan.FromDays(7); // unlimited
-
-						try
-						{
-							using(var cts1 = CancellationTokenSource.CreateLinkedTokenSource(cancellation))
-							{
-								cts1.CancelAfter(loadChainTimeout);
-								node.SynchronizeSlimChain(_Chain, cancellationToken: cts1.Token);
-							}
-						}
-						catch // Timeout happens with SynchronizeChain, if so, throw away the cached chain
-						{
-							_Chain.ResetToGenesis();
-							node.SynchronizeSlimChain(_Chain, cancellationToken: cancellation);
-						}
-
-
-						var peer = (await _RPC.GetPeersInfoAsync())
-									.FirstOrDefault(p => p.SubVersion == userAgent);
-						if(peer != null && !peer.IsWhiteListed)
-						{
-							Logs.Explorer.LogWarning($"{Network.CryptoCode}: Your NBXplorer server is not whitelisted by your node," +
-								$" you should add \"whitelist={peer.Address.Address}\" to the configuration file of your node. (Or use whitebind)");
-						}
-					}
-				}
-				catch(OperationCanceledException) when(!handshaked && handshakeTimeout.IsCancellationRequested)
-				{
-					Logs.Explorer.LogWarning($"{Network.CryptoCode}: The initial hanshake failed, your NBXplorer server might not be whitelisted by your node," +
-							$" if your bitcoin node is on the same machine as NBXplorer, you should add \"whitelist=127.0.0.1\" to the configuration file of your node. (Or use whitebind)");
-					throw;
-				}
-			}
-			Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Height: " + _Chain.Height);
 		}
 
 		private void LoadChainFromCache()
@@ -549,7 +595,7 @@ namespace NBXplorer
 			var suffix = _Network.CryptoCode == "BTC" ? "" : _Network.CryptoCode;
 			{
 				var legacyCachePath = Path.Combine(_Configuration.DataDir, $"{suffix}chain.dat");
-				if(_Configuration.CacheChain && File.Exists(legacyCachePath))
+				if (_Configuration.CacheChain && File.Exists(legacyCachePath))
 				{
 					Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Loading chain from cache...");
 					var chain = new ConcurrentChain(_Network.NBitcoinNetwork);
@@ -563,7 +609,7 @@ namespace NBXplorer
 
 			{
 				var cachePath = Path.Combine(_Configuration.DataDir, $"{suffix}chain-stripped.dat");
-				if(_Configuration.CacheChain && File.Exists(cachePath))
+				if (_Configuration.CacheChain && File.Exists(cachePath))
 				{
 					Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Loading chain from cache...");
 					var chain = new ConcurrentChain(_Network.NBitcoinNetwork);
@@ -581,10 +627,10 @@ namespace NBXplorer
 
 			{
 				var slimCachePath = Path.Combine(_Configuration.DataDir, $"{suffix}chain-slim.dat");
-				if(_Configuration.CacheChain && File.Exists(slimCachePath))
+				if (_Configuration.CacheChain && File.Exists(slimCachePath))
 				{
 					Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Loading chain from cache...");
-					using(var file = new FileStream(slimCachePath, FileMode.Open, FileAccess.Read, FileShare.None, 1024 * 1024))
+					using (var file = new FileStream(slimCachePath, FileMode.Open, FileAccess.Read, FileShare.None, 1024 * 1024))
 					{
 						_Chain.Load(file);
 					}
@@ -596,7 +642,7 @@ namespace NBXplorer
 
 		private void LoadSlimAndSaveToSlimFormat(ConcurrentChain chain)
 		{
-			foreach(var block in chain.ToEnumerable(false))
+			foreach (var block in chain.ToEnumerable(false))
 			{
 				_Chain.TrySetTip(block.HashBlock, block.Previous?.HashBlock);
 			}
@@ -605,29 +651,29 @@ namespace NBXplorer
 
 		private async Task<bool> WarmupBlockchain()
 		{
-			if(await _RPC.GetBlockCountAsync() < _Network.NBitcoinNetwork.Consensus.CoinbaseMaturity)
+			if (await _RPCWithTimeout.GetBlockCountAsync() < _Network.NBitcoinNetwork.Consensus.CoinbaseMaturity)
 			{
 				Logs.Configuration.LogInformation($"{_Network.CryptoCode}: Less than {_Network.NBitcoinNetwork.Consensus.CoinbaseMaturity} blocks, mining some block for regtest");
-				await _RPC.EnsureGenerateAsync(_Network.NBitcoinNetwork.Consensus.CoinbaseMaturity + 1);
+				await _RPCWithTimeout.EnsureGenerateAsync(_Network.NBitcoinNetwork.Consensus.CoinbaseMaturity + 1);
 				return true;
 			}
 			else
 			{
-				var hash = await _RPC.GetBestBlockHashAsync();
+				var hash = await _RPCWithTimeout.GetBestBlockHashAsync();
 
 				BlockHeader header = null;
 				try
 				{
-					header = await _RPC.GetBlockHeaderAsync(hash);
+					header = await _RPCWithTimeout.GetBlockHeaderAsync(hash);
 				}
-				catch(RPCException ex) when(ex.RPCCode == RPCErrorCode.RPC_METHOD_NOT_FOUND)
+				catch (RPCException ex) when (ex.RPCCode == RPCErrorCode.RPC_METHOD_NOT_FOUND)
 				{
-					header = (await _RPC.GetBlockAsync(hash)).Header;
+					header = (await _RPCWithTimeout.GetBlockAsync(hash)).Header;
 				}
-				if((DateTimeOffset.UtcNow - header.BlockTime) > TimeSpan.FromSeconds(24 * 60 * 60))
+				if ((DateTimeOffset.UtcNow - header.BlockTime) > TimeSpan.FromSeconds(24 * 60 * 60))
 				{
 					Logs.Configuration.LogInformation($"{_Network.CryptoCode}: It has been a while nothing got mined on regtest... mining 10 blocks");
-					await _RPC.GenerateAsync(10);
+					await _RPCWithTimeout.GenerateAsync(10);
 					return true;
 				}
 				return false;
@@ -636,13 +682,13 @@ namespace NBXplorer
 
 		public bool IsSynchingCore(GetBlockchainInfoResponse blockchainInfo)
 		{
-			if(blockchainInfo.InitialBlockDownload.HasValue)
-				return blockchainInfo.InitialBlockDownload.Value;
-			if(blockchainInfo.MedianTime.HasValue && _Network.NBitcoinNetwork.NetworkType != NetworkType.Regtest)
+			if (blockchainInfo.InitialBlockDownload == true)
+				return true;
+			if (blockchainInfo.MedianTime.HasValue && _Network.NBitcoinNetwork.NetworkType != NetworkType.Regtest)
 			{
 				var time = NBitcoin.Utils.UnixTimeToDateTime(blockchainInfo.MedianTime.Value);
 				// 5 month diff? probably synching...
-				if(DateTimeOffset.UtcNow - time > TimeSpan.FromDays(30 * 5))
+				if (DateTimeOffset.UtcNow - time > TimeSpan.FromDays(30 * 5))
 				{
 					return true;
 				}
@@ -657,9 +703,11 @@ namespace NBXplorer
 		{
 			get
 			{
-				return _Group?.ConnectedNodes.Count != 0;
+				return GetHandshakedNode() != null;
 			}
 		}
 
+		public GetNetworkInfoResponse NetworkInfo { get; internal set; }
+		public bool HasTxIndex { get; set; }
 	}
 }
