@@ -27,7 +27,7 @@ namespace NBXplorer.Configuration
 			get;
 			internal set;
 		}
-		public IPEndPoint NodeEndpoint
+		public EndPoint NodeEndpoint
 		{
 			get;
 			internal set;
@@ -37,11 +37,17 @@ namespace NBXplorer.Configuration
 			get;
 			internal set;
 		}
+		public Money MinUtxoValue
+		{
+			get;
+			internal set;
+		}
 		public string CryptoCode
 		{
 			get;
 			set;
 		}
+		public bool HasTxIndex { get; set; }
 	}
 	public class ExplorerConfiguration
 	{
@@ -60,7 +66,7 @@ namespace NBXplorer.Configuration
 		{
 			get; set;
 		}
-
+		public string SignalFilesDir { get; set; }
 		public NBXplorerNetworkProvider NetworkProvider
 		{
 			get; set;
@@ -70,7 +76,6 @@ namespace NBXplorer.Configuration
 			get;
 			set;
 		}
-
 		public int MinGapSize
 		{
 			get; set;
@@ -115,15 +120,36 @@ namespace NBXplorer.Configuration
 					chainConfiguration.CryptoCode = network.CryptoCode;
 
 					var args = RPCArgs.Parse(config, network.NBitcoinNetwork, network.CryptoCode);
+
 					chainConfiguration.RPC = args.ConfigureRPCClient(network);
+					if (chainConfiguration.RPC.Address.Port == network.NBitcoinNetwork.DefaultPort)
+					{
+						Logs.Configuration.LogWarning($"{network.CryptoCode}: It seems that the RPC port ({chainConfiguration.RPC.Address.Port}) is equal to the default P2P port ({network.NBitcoinNetwork.DefaultPort}), this is probably a misconfiguration.");
+					}
 					if((chainConfiguration.RPC.CredentialString.CookieFile != null || chainConfiguration.RPC.CredentialString.UseDefault) && !network.SupportCookieAuthentication)
 					{
 						throw new ConfigException($"Chain {network.CryptoCode} does not support cookie file authentication,\n" +
 							$"Please use {network.CryptoCode.ToLowerInvariant()}rpcuser and {network.CryptoCode.ToLowerInvariant()}rpcpassword settings in NBXplorer" +
 							$"And configure rpcuser and rpcpassword in the configuration file or in commandline or your node");
 					}
-					chainConfiguration.NodeEndpoint = DefaultConfiguration.ConvertToEndpoint(config.GetOrDefault<string>($"{network.CryptoCode}.node.endpoint", "127.0.0.1"), network.NBitcoinNetwork.DefaultPort);
+					chainConfiguration.NodeEndpoint = NBitcoin.Utils.ParseEndpoint(config.GetOrDefault<string>($"{network.CryptoCode}.node.endpoint", "127.0.0.1"), network.NBitcoinNetwork.DefaultPort);
+
+					if (GetPort(chainConfiguration.NodeEndpoint) == network.NBitcoinNetwork.RPCPort)
+					{
+						Logs.Configuration.LogWarning($"{network.CryptoCode}: It seems that the node endpoint port ({GetPort(chainConfiguration.NodeEndpoint)}) is equal to the default RPC port ({network.NBitcoinNetwork.RPCPort}), this is probably a misconfiguration.");
+					}
+
 					chainConfiguration.StartHeight = config.GetOrDefault<int>($"{network.CryptoCode}.startheight", -1);
+
+					if (!(network is NBXplorer.NBXplorerNetworkProvider.LiquidNBXplorerNetwork))
+					{
+						if (config.GetOrDefault<int>($"{network.CryptoCode}.minutxovalue", -1) is int v && v != -1)
+						{
+							chainConfiguration.MinUtxoValue = Money.Satoshis(v);
+						}
+					}
+					
+					chainConfiguration.HasTxIndex = config.GetOrDefault<bool>($"{network.CryptoCode}.hastxindex", false);
 
 					ChainConfigurations.Add(chainConfiguration);
 				}
@@ -140,10 +166,24 @@ namespace NBXplorer.Configuration
 			if(!Directory.Exists(BaseDataDir))
 				Directory.CreateDirectory(BaseDataDir);
 			DataDir = Path.Combine(BaseDataDir, NBXplorerDefaultSettings.GetFolderName(NetworkProvider.NetworkType));
-			if(!Directory.Exists(DataDir))
+			if (!Directory.Exists(DataDir))
 				Directory.CreateDirectory(DataDir);
+			SignalFilesDir = config.GetOrDefault<string>("signalfilesdir", null);
+			SignalFilesDir = SignalFilesDir ?? DataDir;
+			if (!Directory.Exists(SignalFilesDir))
+				Directory.CreateDirectory(SignalFilesDir);
 			CacheChain = config.GetOrDefault<bool>("cachechain", true);
 			NoAuthentication = config.GetOrDefault<bool>("noauth", false);
+
+			var customKeyPathTemplate = config.GetOrDefault<string>("customkeypathtemplate", null);
+			if (!string.IsNullOrEmpty(customKeyPathTemplate))
+			{
+				if (!KeyPathTemplate.TryParse(customKeyPathTemplate, out var v))
+					throw new ConfigException("Invalid customKeyPathTemplate");
+				if (v.PostIndexes.IsHardened || v.PreIndexes.IsHardened)
+					throw new ConfigException("customKeyPathTemplate should not be an hardened path");
+				CustomKeyPathTemplate = v;
+			}
 
 			AzureServiceBusConnectionString = config.GetOrDefault<string>("asbcnstr", "");
 			AzureServiceBusBlockQueue = config.GetOrDefault<string>("asbblockq", "");
@@ -151,7 +191,23 @@ namespace NBXplorer.Configuration
 			AzureServiceBusBlockTopic = config.GetOrDefault<string>("asbblockt", "");
 			AzureServiceBusTransactionTopic = config.GetOrDefault<string>("asbtrant", "");
 
+			RabbitMqHostName = config.GetOrDefault<string>("rmqhost", "");
+			RabbitMqVirtualHost = config.GetOrDefault<string>("rmqvirtual", "");
+			RabbitMqUsername = config.GetOrDefault<string>("rmquser", "");
+			RabbitMqPassword = config.GetOrDefault<string>("rmqpass", "");
+			RabbitMqTransactionExchange = config.GetOrDefault<string>("rmqtranex", "");
+			RabbitMqBlockExchange = config.GetOrDefault<string>("rmqblockex", "");
+
 			return this;
+		}
+
+		private int GetPort(EndPoint nodeEndpoint)
+		{
+			if (nodeEndpoint is IPEndPoint endPoint)
+				return endPoint.Port;
+			else if (nodeEndpoint is DnsEndPoint dnsEndPoint)
+				return dnsEndPoint.Port;
+			throw new NotSupportedException();
 		}
 
 		public bool Supports(NBXplorerNetwork network)
@@ -197,5 +253,14 @@ namespace NBXplorer.Configuration
 			get;
 			set;
 		}
-	}
+
+		public string RabbitMqHostName { get; set; }
+        public string RabbitMqVirtualHost { get; set; }
+        public string RabbitMqUsername { get; set; }
+        public string RabbitMqPassword { get; set; }
+        public string RabbitMqTransactionExchange { get; set; }
+        public string RabbitMqBlockExchange { get; set; }
+
+		public KeyPathTemplate CustomKeyPathTemplate { get; set; }
+    }
 }
