@@ -35,6 +35,7 @@ namespace NBXplorer
 			_EventAggregator = eventAggregator;
 		}
 
+		CancellationTokenSource _Cts = new CancellationTokenSource();
 		EventAggregator _EventAggregator;
 
 		Repository Repository
@@ -68,25 +69,24 @@ namespace NBXplorer
 
 		Timer _Timer;
 
-		public BlockLocator CurrentLocation
-		{
-			get
-			{
-				return _CurrentLocation;
-			}
-		}
+		public BlockLocator CurrentLocation { get; private set; }
 
 		protected override void AttachCore()
 		{
 			AttachedNode.StateChanged += AttachedNode_StateChanged;
 			AttachedNode.MessageReceived += AttachedNode_MessageReceived;
-			_CurrentLocation = Repository.GetIndexProgress() ?? GetDefaultCurrentLocation();
-			var fork = Chain.FindFork(_CurrentLocation);
+		}
+
+		public async Task Init()
+		{
+			var currentLocation = await Repository.GetIndexProgress() ?? GetDefaultCurrentLocation();
+			var fork = Chain.FindFork(currentLocation);
 			if (fork == null)
 			{
-				_CurrentLocation = GetDefaultCurrentLocation();
-				fork = Chain.FindFork(_CurrentLocation);
+				currentLocation = GetDefaultCurrentLocation();
+				fork = Chain.FindFork(currentLocation);
 			}
+			CurrentLocation = currentLocation;
 			Logs.Explorer.LogInformation($"{Network.CryptoCode}: Starting scan at block " + fork.Height);
 			_Timer = new Timer(Tick, null, 0, (int)TimeSpan.FromSeconds(30).TotalMilliseconds);
 		}
@@ -95,58 +95,80 @@ namespace NBXplorer
 		{
 			if (StartHeight > Chain.Height)
 				throw new InvalidOperationException($"{Network.CryptoCode}: StartHeight should not be above the current tip");
-			return StartHeight == -1 ?
-				Chain.GetTipLocator() :
-				Chain.GetLocator(StartHeight);
+
+			BlockLocator blockLocator = null;
+			if (StartHeight == -1)
+			{
+				blockLocator = Chain.GetTipLocator();
+				Logs.Explorer.LogInformation($"{Network.CryptoCode}: Current Index Progress not found, start syncing from the header's chain tip (At height: {Chain.Height})");
+			}
+			else
+			{
+				blockLocator = Chain.GetLocator(StartHeight);
+				Logs.Explorer.LogInformation($"{Network.CryptoCode}: Current Index Progress not found, start syncing at height {Chain.Height}");
+			}
+			return blockLocator;
 		}
 
 
 
-		public void AskBlocks()
+		public int AskBlocks()
 		{
 			var node = AttachedNode;
 			if (node == null || node.State != NodeState.HandShaked)
-				return;
+				return 0;
 			if (Chain.Height < node.PeerVersion.StartHeight)
-				return;
-			if (_InFlights.Count != 0)
-				return;
-			var currentLocation = _CurrentLocation;
+				return 0;
+			var currentLocation = (_HighestInFlight ?? CurrentLocation);
+			if (currentLocation == null)
+				return 0;
 			var currentBlock = Chain.FindFork(currentLocation);
 			if (currentBlock.Height < StartHeight)
 				currentBlock = Chain.GetBlock(StartHeight) ?? Chain.TipBlock;
 
 			//Up to date
 			if (Chain.TipBlock.Hash == currentBlock.Hash)
-				return;
+				return 0;
 
 
-			var invs = Enumerable.Range(0, 50)
-				.Select(i => Chain.GetBlock(i + currentBlock.Height + 1))
-				.Where(_ => _ != null)
-				.Take(40)
-				.Select(b => new InventoryVector(node.AddSupportedOptions(InventoryType.MSG_BLOCK), b.Hash))
-				.Where(b => _InFlights.TryAdd(b.Hash, new Download()))
-				.ToArray();
-
-			if (invs.Length != 0)
+			int maxConcurrentBlocks = 10;
+			List<InventoryVector> invs = new List<InventoryVector>();
+			foreach (var i in Enumerable.Range(0, int.MaxValue))
 			{
-				node.SendMessageAsync(new GetDataPayload(invs));
+				var block = Chain.GetBlock(currentBlock.Height + 1 + i);
+				if (block == null)
+					break;
+				if (_InFlights.TryAdd(block.Hash, block.Hash))
+					invs.Add(new InventoryVector(node.AddSupportedOptions(InventoryType.MSG_BLOCK), block.Hash));
+				if (invs.Count == maxConcurrentBlocks)
+					break;
 			}
+
+			if (invs.Count != 0)
+			{
+				Repository.BatchSize = invs.Count == maxConcurrentBlocks ? int.MaxValue : 100;
+				_HighestInFlight = _HighestInFlight ?? Chain.GetLocator(invs[invs.Count - 1].Hash);
+				node.SendMessageAsync(new GetDataPayload(invs.ToArray()));
+				if (invs.Count > 1)
+					GC.Collect(); // Let's collect memory if we are synching 
+			}
+			return invs.Count;
 		}
 
-		class Download
-		{
-		}
-
-		ConcurrentDictionary<uint256, Download> _InFlights = new ConcurrentDictionary<uint256, Download>();
-
-
+		ConcurrentDictionary<uint256, uint256> _InFlights = new ConcurrentDictionary<uint256, uint256>();
+		BlockLocator _HighestInFlight;
+		DateTimeOffset? _LastBlockDownloaded;
+		readonly static TimeSpan DownloadHangingTimeout = TimeSpan.FromMinutes(1.0);
 		void Tick(object state)
 		{
 			try
 			{
-				AskBlocks();
+				if (AskBlocks() == 0 &&
+					IsHanging())
+				{
+					Logs.Explorer.LogInformation($"{Network.CryptoCode}: Block download seems to hang, let's reconnect to this node");
+					AttachedNode?.DisconnectAsync("Block download is hanging");
+				}
 			}
 			catch (Exception ex)
 			{
@@ -155,6 +177,32 @@ namespace NBXplorer
 				Logs.Explorer.LogError($"{Network.CryptoCode}: Exception in ExplorerBehavior tick loop");
 				Logs.Explorer.LogError(ex.ToString());
 			}
+		}
+
+		public bool IsDownloading => !_InFlights.IsEmpty;
+
+		public bool IsHanging()
+		{
+			var node = AttachedNode;
+			if (node == null)
+				return false;
+			DateTimeOffset lastActivity;
+			if (IsDownloading) // If we download
+			{
+				if (_LastBlockDownloaded is DateTimeOffset lastBlockDownloaded)
+					lastActivity = lastBlockDownloaded;
+				else
+					lastActivity = node.ConnectedAt;
+			}
+			else if (IsSynching()) // Not downloading but synching
+			{
+				lastActivity = node.ConnectedAt;
+			}
+			else // If we do not download and we do not sync, we are ok
+			{
+				return false;
+			}
+			return (DateTimeOffset.UtcNow - lastActivity) >= DownloadHangingTimeout;
 		}
 
 		public NBXplorerNetwork Network
@@ -170,21 +218,21 @@ namespace NBXplorer
 			get;
 		}
 
-		BlockLocator _CurrentLocation;
-
 		protected override void DetachCore()
 		{
 			AttachedNode.StateChanged -= AttachedNode_StateChanged;
 			AttachedNode.MessageReceived -= AttachedNode_MessageReceived;
-
+			_Cts.Cancel();
 			_Timer.Dispose();
 			_Timer = null;
 		}
-
 		private void AttachedNode_MessageReceived(Node node, IncomingMessage message)
 		{
-			message.Message.IfPayloadIs<InvPayload>(invs =>
+			if (message.Message.Payload is InvPayload invs)
 			{
+				// Do not asks transactions if we are synching so that we can process blocks faster
+				if (IsSynching())
+					return;
 				var data = new GetDataPayload();
 				foreach (var inv in invs.Inventory)
 				{
@@ -194,76 +242,147 @@ namespace NBXplorer
 				}
 				if (data.Inventory.Count != 0)
 					node.SendMessageAsync(data);
-			});
-
-			message.Message.IfPayloadIs<HeadersPayload>(headers =>
+			}
+			else if (message.Message.Payload is HeadersPayload headers)
 			{
 				if (headers.Headers.Count == 0)
 					return;
 				AskBlocks();
-			});
-
-			message.Message.IfPayloadIs<BlockPayload>(block =>
+			}
+			else if (message.Message.Payload is BlockPayload block)
 			{
-				block.Object.Header.PrecomputeHash(false, false);
-				Download o;
-				if (_InFlights.ContainsKey(block.Object.GetHash()))
+				Run(() => SaveMatches(block.Object));
+			}
+			else if (message.Message.Payload is TxPayload txPayload)
+			{
+				Run(() => SaveMatches(txPayload.Object));
+			}
+		}
+		Task Run(Func<Task> act)
+		{
+			return Task.Run(async () =>
+			{
+				try
 				{
-					var currentLocation = Chain.GetLocator(block.Object.GetHash());
-					if (currentLocation == null)
-						return;
-					_CurrentLocation = currentLocation;
-					if (_InFlights.TryRemove(block.Object.GetHash(), out o))
-					{
-						foreach (var tx in block.Object.Transactions)
-							tx.PrecomputeHash(false, true);
-
-						var matches =
-							block.Object.Transactions
-							.SelectMany(tx => Repository.GetMatches(tx))
-							.ToArray();
-
-						var blockHash = block.Object.GetHash();
-						SaveMatches(matches, blockHash);
-						//Save index progress everytimes if not synching, or once every 100 blocks otherwise
-						if (!IsSynching() || blockHash.GetLow32() % 100 == 0)
-							Repository.SetIndexProgress(currentLocation);
-						_EventAggregator.Publish(new Events.NewBlockEvent(this._Repository.Network.CryptoCode, blockHash));
-					}
-					if (_InFlights.Count == 0)
-						AskBlocks();
+					await act();
+				}
+				catch (Exception ex)
+				{
+					Logs.Explorer.LogError($"{Network.CryptoCode}: Unhandled error while treating a message");
+					Logs.Explorer.LogError(ex.ToString());
+					this.AttachedNode.DisconnectAsync($"{Network.CryptoCode}: Unhandled error while treating a message", ex);
 				}
 			});
-
-			message.Message.IfPayloadIs<TxPayload>(txPayload =>
-			{
-				var matches = Repository.GetMatches(txPayload.Object).ToArray();
-				SaveMatches(matches, null);
-			});
-
 		}
-
-		private void SaveMatches(TransactionMatch[] matches, uint256 blockHash)
+		private async Task SaveMatches(Block block)
 		{
-			DateTimeOffset now = DateTimeOffset.UtcNow;
-			var matchedTransactions = matches.Select(m => new MatchedTransaction()
+			_LastBlockDownloaded = DateTimeOffset.UtcNow;
+			block.Header.PrecomputeHash(false, false);
+			foreach (var tx in block.Transactions)
+				tx.PrecomputeHash(false, true);
+
+			var blockHash = block.GetHash();
+			var delay = TimeSpan.FromSeconds(1);
+		retry:
+			try
 			{
-				BlockId = blockHash,
-				Match = m,
-			}).ToArray();
-			Repository.SaveMatches(now, matchedTransactions);
-			AddressPoolService.RefillAddressPoolIfNeeded(Network, matchedTransactions);
-			var saved = Repository.SaveTransactions(now, matches.Select(m => m.Transaction).Distinct().ToArray(), blockHash);
-			var savedTransactions = saved.ToDictionary(s => s.Transaction.GetHash());
-			for (int i = 0; i < matches.Length; i++)
+				DateTimeOffset now = DateTimeOffset.UtcNow;
+				var matches =
+					(await Repository.GetMatches(block.Transactions, blockHash, now, true))
+					.ToArray();
+				await SaveMatches(matches, blockHash, now);
+				var slimBlockHeader = Chain.GetBlock(blockHash);
+				if (slimBlockHeader != null)
+				{
+					var blockEvent = new Models.NewBlockEvent()
+					{
+						CryptoCode = _Repository.Network.CryptoCode,
+						Hash = blockHash,
+						Height = slimBlockHeader.Height,
+						PreviousBlockHash = slimBlockHeader.Previous
+					};
+					var saving = Repository.SaveEvent(blockEvent);
+					_EventAggregator.Publish(blockEvent);
+					await saving;
+				}
+			}
+			catch (ObjectDisposedException)
 			{
-				_EventAggregator.Publish(new NewTransactionMatchEvent(this._Repository.Network.CryptoCode, blockHash, matches[i], savedTransactions[matches[i].Transaction.GetHash()]));
+				throw;
+			}
+			catch (Exception ex)
+			{
+				Logs.Explorer.LogWarning(ex, $"{Network.CryptoCode}: Error while saving block in database, retrying in {delay.TotalSeconds} seconds ({ex.Message})");
+				await Task.Delay(delay, _Cts.Token);
+				delay = delay * 2;
+				var maxDelay = TimeSpan.FromSeconds(60);
+				if (delay > maxDelay)
+					delay = maxDelay;
+				goto retry;
+			}
+
+			if (_InFlights.TryRemove(blockHash, out var unused) && _InFlights.IsEmpty)
+			{
+				var highestInFlight = _HighestInFlight;
+				_HighestInFlight = null;
+				if (highestInFlight != null)
+				{
+					CurrentLocation = highestInFlight;
+					await Repository.SetIndexProgress(highestInFlight);
+				}
+				AskBlocks();
 			}
 		}
 
+
+		private async Task SaveMatches(Transaction transaction)
+		{
+			var now = DateTimeOffset.UtcNow;
+			var matches = (await Repository.GetMatches(transaction, null, now, false)).ToArray();
+			await SaveMatches(matches, null, now);
+		}
+
+		private async Task SaveMatches(TrackedTransaction[] matches, uint256 blockHash, DateTimeOffset now)
+		{
+			await Repository.SaveMatches(matches);
+			_ = AddressPoolService.GenerateAddresses(Network, matches);
+			var saved = await Repository.SaveTransactions(now, matches.Select(m => m.Transaction).Distinct().ToArray(), blockHash);
+			var savedTransactions = saved.ToDictionary(s => s.Transaction.GetHash());
+
+			int? maybeHeight = null;
+			var chainHeight = Chain.Height;
+			if (blockHash != null && Chain.TryGetHeight(blockHash, out int height))
+			{
+				maybeHeight = height;
+			}
+			Task[] saving = new Task[matches.Length];
+			for (int i = 0; i < matches.Length; i++)
+			{
+				var txEvt = new Models.NewTransactionEvent()
+				{
+					TrackedSource = matches[i].TrackedSource,
+					DerivationStrategy = (matches[i].TrackedSource is DerivationSchemeTrackedSource dsts) ? dsts.DerivationStrategy : null,
+					CryptoCode = Network.CryptoCode,
+					BlockId = blockHash,
+					TransactionData = new TransactionResult()
+					{
+						BlockId = blockHash,
+						Height = maybeHeight,
+						Confirmations = maybeHeight == null ? 0 : chainHeight - maybeHeight.Value + 1,
+						Timestamp = now,
+						Transaction = matches[i].Transaction,
+						TransactionHash = matches[i].TransactionHash
+					},
+					Outputs = matches[i].GetReceivedOutputs().ToList()
+				};
+				saving[i] = Repository.SaveEvent(txEvt);
+				_EventAggregator.Publish(txEvt);
+			}
+			await Task.WhenAll(saving);
+		}
 		public bool IsSynching()
 		{
-			var location = _CurrentLocation;
+			var location = CurrentLocation;
 			if (location == null)
 				return true;
 			var fork = Chain.FindFork(location);
