@@ -7,6 +7,7 @@ using Microsoft.Extensions.Options;
 using NBXplorer.Configuration;
 using NBXplorer.Logging;
 using Newtonsoft.Json;
+using RabbitMQ.Client;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -23,78 +24,61 @@ namespace NBXplorer.MessageBrokers
 		IBrokerClient _senderBlock = null;
 		IBrokerClient _senderTransactions = null;
 		ExplorerConfiguration _config;
-		JsonSerializerSettings _serializerSettings;
 
-		public BrokerHostedService(BitcoinDWaitersAccessor waiters, ChainProvider chainProvider, EventAggregator eventAggregator, IOptions<ExplorerConfiguration> config, IOptions<MvcJsonOptions> jsonOptions)
+		public BrokerHostedService(BitcoinDWaiters waiters, ChainProvider chainProvider, EventAggregator eventAggregator, IOptions<ExplorerConfiguration> config, NBXplorerNetworkProvider networks)
 		{
 			_EventAggregator = eventAggregator;
+			Networks = networks;
 			ChainProvider = chainProvider;
-			Waiters = waiters.Instance;
+			Waiters = waiters;
 			_config = config.Value;
-			_serializerSettings = jsonOptions.Value.SerializerSettings;
 		}
 
 		public Task StartAsync(CancellationToken cancellationToken)
 		{
-			if(_Disposed)
+			if (_Disposed)
 				throw new ObjectDisposedException(nameof(BrokerHostedService));
 
 			_senderBlock = CreateClientBlock();
 			_senderTransactions = CreateClientTransaction();
 
-			_subscriptions.Add(_EventAggregator.Subscribe<Events.NewBlockEvent>(async o =>
+			_subscriptions.Add(_EventAggregator.Subscribe<Models.NewBlockEvent>(async o =>
 			{
-				var chain = ChainProvider.GetChain(o.CryptoCode);
-				if(chain == null)
-					return;
-				var block = chain.GetBlock(o.BlockId);
-				if(block != null)
-				{
-					var nbe = new Models.NewBlockEvent()
-					{
-						CryptoCode = o.CryptoCode,
-						Hash = block.Hash,
-						Height = block.Height,
-						PreviousBlockHash = block?.Previous
-					};
-					await _senderBlock.Send(nbe);
-				}
+				await _senderBlock.Send(o);
 			}));
 
 
-			_subscriptions.Add(_EventAggregator.Subscribe<Events.NewTransactionMatchEvent>(async o =>
+			_subscriptions.Add(_EventAggregator.Subscribe<Models.NewTransactionEvent>(async o =>
 			{
-				var network = Waiters.GetWaiter(o.CryptoCode);
-				if(network == null)
-					return;
-				var chain = ChainProvider.GetChain(o.CryptoCode);
-				if(chain == null)
-					return;
-				var txe = new Models.NewTransactionEvent()
-				{
-					TrackedSource = o.Match.TrackedSource,
-					DerivationStrategy = o.Match.DerivationStrategy,
-					CryptoCode = o.CryptoCode,
-					BlockId = o.BlockId,
-					TransactionData = Utils.ToTransactionResult(true, chain, new[] { o.SavedTransaction }),
-					Inputs = o.Match.Inputs,
-					Outputs = o.Match.Outputs
-				};
-				await _senderTransactions.Send(txe);
+				await _senderTransactions.Send(o);
 			}));
-
 			return Task.CompletedTask;
 		}
 
 		IBrokerClient CreateClientTransaction()
 		{
 			var brokers = new List<IBrokerClient>();
-			if(!string.IsNullOrEmpty(_config.AzureServiceBusConnectionString))
+			if (!string.IsNullOrEmpty(_config.AzureServiceBusConnectionString))
 			{
-				if(!string.IsNullOrWhiteSpace(_config.AzureServiceBusTransactionQueue))
+				if (!string.IsNullOrWhiteSpace(_config.AzureServiceBusTransactionQueue))
 					brokers.Add(CreateAzureQueue(_config.AzureServiceBusConnectionString, _config.AzureServiceBusTransactionQueue));
-				if(!string.IsNullOrWhiteSpace(_config.AzureServiceBusTransactionTopic))
+				if (!string.IsNullOrWhiteSpace(_config.AzureServiceBusTransactionTopic))
 					brokers.Add(CreateAzureTopic(_config.AzureServiceBusConnectionString, _config.AzureServiceBusTransactionTopic));
+			}
+			if(!string.IsNullOrEmpty(_config.RabbitMqHostName) && 
+				!string.IsNullOrEmpty(_config.RabbitMqUsername) && 
+				!string.IsNullOrEmpty(_config.RabbitMqPassword)) 
+			{
+				if(!string.IsNullOrEmpty(_config.RabbitMqTransactionExchange)) 
+				{
+					brokers.Add(CreateRabbitMqExchange(
+						hostName: _config.RabbitMqHostName,
+						virtualHost: _config.RabbitMqVirtualHost,
+						username: _config.RabbitMqUsername,
+						password: _config.RabbitMqPassword,
+						newTransactionExchange: _config.RabbitMqTransactionExchange,
+						newBlockExchange: string.Empty));
+				}
 			}
 			return new CompositeBroker(brokers);
 		}
@@ -102,24 +86,53 @@ namespace NBXplorer.MessageBrokers
 		IBrokerClient CreateClientBlock()
 		{
 			var brokers = new List<IBrokerClient>();
-			if(!string.IsNullOrEmpty(_config.AzureServiceBusConnectionString))
+			if (!string.IsNullOrEmpty(_config.AzureServiceBusConnectionString))
 			{
-				if(!string.IsNullOrWhiteSpace(_config.AzureServiceBusBlockQueue))
+				if (!string.IsNullOrWhiteSpace(_config.AzureServiceBusBlockQueue))
 					brokers.Add(CreateAzureQueue(_config.AzureServiceBusConnectionString, _config.AzureServiceBusBlockQueue));
-				if(!string.IsNullOrWhiteSpace(_config.AzureServiceBusBlockTopic))
+				if (!string.IsNullOrWhiteSpace(_config.AzureServiceBusBlockTopic))
 					brokers.Add(CreateAzureTopic(_config.AzureServiceBusConnectionString, _config.AzureServiceBusBlockTopic));
+			}
+
+			if(!string.IsNullOrEmpty(_config.RabbitMqHostName) && 
+				!string.IsNullOrEmpty(_config.RabbitMqUsername) && 
+				!string.IsNullOrEmpty(_config.RabbitMqPassword)) 
+			{
+				if(!string.IsNullOrEmpty(_config.RabbitMqBlockExchange)) 
+				{
+					brokers.Add(CreateRabbitMqExchange(
+						hostName: _config.RabbitMqHostName,
+						virtualHost: _config.RabbitMqVirtualHost,
+						username: _config.RabbitMqUsername,
+						password: _config.RabbitMqPassword,
+						newTransactionExchange: string.Empty,
+						newBlockExchange: _config.RabbitMqBlockExchange));
+				}
 			}
 			return new CompositeBroker(brokers);
 		}
 
 		private IBrokerClient CreateAzureQueue(string connnectionString, string queueName)
 		{
-			return new AzureBroker(new QueueClient(connnectionString, queueName), _serializerSettings);
+			return new AzureBroker(new QueueClient(connnectionString, queueName), Networks);
 		}
 
 		private IBrokerClient CreateAzureTopic(string connectionString, string topicName)
 		{
-			return new AzureBroker(new TopicClient(connectionString, topicName), _serializerSettings);
+			return new AzureBroker(new TopicClient(connectionString, topicName), Networks);
+		}
+
+		private IBrokerClient CreateRabbitMqExchange(
+			string hostName, string virtualHost, 
+			string username, string password, 
+			string newTransactionExchange, string newBlockExchange)
+		{
+			return new RabbitMqBroker(
+				Networks,
+				new ConnectionFactory() { 
+					HostName = hostName, VirtualHost = virtualHost,
+					UserName = username, Password = password }, 
+				newTransactionExchange, newBlockExchange );
 		}
 
 		public async Task StopAsync(CancellationToken cancellationToken)
@@ -133,6 +146,7 @@ namespace NBXplorer.MessageBrokers
 		{
 			get; set;
 		}
+		public NBXplorerNetworkProvider Networks { get; }
 		public BitcoinDWaiters Waiters
 		{
 			get; set;
