@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Mvc.Internal;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
+﻿using Microsoft.AspNetCore.Mvc.ModelBinding;
 using NBitcoin;
 using System;
 using System.Reflection;
