@@ -3,6 +3,7 @@ using NBXplorer.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net.Http;
 using System.Threading.Tasks;
 
 namespace NBXplorer
@@ -10,13 +11,16 @@ namespace NBXplorer
 	public class RPCClientProvider
 	{
 		Dictionary<string, RPCClient> _ChainConfigurations = new Dictionary<string, RPCClient>();
-		public RPCClientProvider(ExplorerConfiguration configuration)
+		public RPCClientProvider(ExplorerConfiguration configuration, IHttpClientFactory httpClientFactory)
 		{
 			foreach(var config in configuration.ChainConfigurations)
 			{
 				var rpc = config?.RPC;
-				if(rpc != null)
+				if (rpc != null)
+				{
+					rpc.HttpClient = httpClientFactory.CreateClient(nameof(RPCClientProvider));
 					_ChainConfigurations.Add(config.CryptoCode, rpc);
+				}
 			}
 		}
 
